% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_archive.R
\name{cas_archive}
\alias{cas_archive}
\title{Archive originals of downloaded files in compressed folders}
\usage{
cas_archive(
  path = NULL,
  file_format = "tar.gz",
  index = TRUE,
  contents = TRUE,
  remove_original = TRUE,
  db_connection = NULL,
  db_folder = NULL,
  ...
)
}
\arguments{
\item{path}{Path to archive directory, defaults to NULL. If NULL, path is set
to the project/website/archive folder.}

\item{file_format}{Defaults to "tar.gz", to ensure cross-platform
compatibility. No other formats are supported at this stage.}

\item{remove_original}{Defaults to TRUE. If TRUE, after local files have been
confirmed to be stored in the relevant compressed file, they are removed
from their original folders, and the empty folders deleted.}

\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\value{
Nothing, used for its side effects.
}
\description{
Archive originals of downloaded files in compressed folders
}
