% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_backup.R
\name{cas_backup_gd}
\alias{cas_backup_gd}
\title{Backup files to Google Drive}
\usage{
cas_backup_gd(
  glob = c("*.tar.gz", "*.sqlite"),
  email = gargle::gargle_oauth_email(),
  scopes = "https://www.googleapis.com/auth/drive.file",
  client = cas_google_client,
  ...
)
}
\arguments{
\item{glob}{A character vector with all glob selectors for the type of files
to be stored. Defaults to \code{c("*.tar.gz", "*.sqlite")}.}

\item{email}{If given, email of the Google account to use for storing files.}

\item{scopes}{Defaults to \code{drive.file}, i.e., only give access to files
created with this client. This means that no access to other files and
folders on your Google Drive is ever given to this session.}

\item{client}{Google app client, defaults to \code{castarter}'s own. Passed to
\code{googledrive::drive_auth_configure}. Set to NULL to use \code{googledrive}'s
defaults.}
}
\description{
Backup files to Google Drive
}
