% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_build_urls.R
\name{cas_build_urls}
\alias{cas_build_urls}
\title{Generates urls to index pages listing individual articles..}
\usage{
cas_build_urls(
  url_first_part,
  url_second_part = NULL,
  start_page = 1,
  end_page = 10,
  increase_by = 1,
  date_format = "YMD",
  start_date = NULL,
  end_date = NULL,
  date_separator = NULL,
  reversed_order = FALSE,
  project = NULL,
  website = NULL,
  cache = NULL,
  cache_connection = NULL
)
}
\arguments{
\item{url_first_part}{First part of index link that does not change in other
index pages.}

\item{url_second_part}{Part of index link appneded after the part of the link
that varies. If not relevant, may be left empty.}

\item{start_page}{If the urls include a numerical component, define first
number of the sequence. start_page defaults to 1.}

\item{end_page}{If the urls include a numerical component, define first
number of the sequence. end_page defaults to 10.}

\item{increase_by}{Defines by how much the number in the link should be
increased in the numerical sequence. Defaults to 1.}

\item{date_format}{A character string, defaults to "YMD". Check
\code{\link{strptime}} for valid values used to define the format of the
date that is part of the URL. Simplified formats such as the following are
also accepted: "Y" (e.g. 2022), "Ym" (2022-10), "Ymd" (e.g. 2022-10-24).
Read more in details.}

\item{start_date, end_date}{Defaults to NULL. If given, a date, or a character
vector coercible to date with \code{\link{as.Date}}.}

\item{project}{Name of 'castarter' project. Must correspond to the name of a
folder in the current working directory. Not required if previously set
with SetCastarter(project = "project", website = "website")}

\item{website}{Name of a website included in a 'castarter' project. Must
correspond to the name of a sub-folder of the project folder. Not required
if previously set with SetCastarter(project = "project", website =
"website")}
}
\value{
A character vector of urls to index pages.
}
\description{
Generates urls to index pages listing individual articles.
}
\section{Date formats}{
 It is not uncommon in particular for index pages to
include dates in the URL, along the lines of
\code{example.com/archive/2022-01-01}, \code{example.com/archive/2022-01-02}, etc. To
build such urls, \code{cas_build_urls} needs a \code{start_date} and \code{end_date}.
The formatting of the date can be defined either by providing to the
parameter \code{date_format} a string that \code{\link{strptime}} is able to
interpret directly, or a simplified string (such as "YMD", without the
"\%"), and add a \code{date_separator} such as "-" as needed.
}

\examples{
\dontrun{
index_urls <- cas_build_urls("http://www.example.com/news/")
}
}
