% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_check_read_db_contents_data.R
\name{cas_check_read_db_contents_data}
\alias{cas_check_read_db_contents_data}
\title{Returns a corpus from the \code{contents_data} table in the database; if corpus is give, it just returns that instead.}
\usage{
cas_check_read_db_contents_data(
  corpus = NULL,
  collect = FALSE,
  db_connection = NULL,
  db_folder = NULL,
  ...
)
}
\arguments{
\item{collect}{Logical, defaults to FALSE. If TRUE, it always returns a data
frame and not a database connection, no matter the input.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\description{
Mostly used internally
}
