% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_files.R
\name{cas_check_website_folder}
\alias{cas_check_website_folder}
\title{Checks if current website folder exists}
\usage{
cas_check_website_folder(base_folder = NULL, project = NULL, website = NULL)
}
\arguments{
\item{base_folder}{Defaults to NULL, can be set once per session with
\code{\link[=cas_set_options]{cas_set_options()}}. A path to a location used for storing html and other
project files. If the folder does not exist, it will be created. If not
given, and not previously set as environment variable, defaults to
\code{castarter_data}.}

\item{project}{Defaults to NULL. Project name, can be set once per session
with \code{\link[=cas_set_options]{cas_set_options()}}. This will be used as first level folder and may
be used elsewhere to describe the dataset.}

\item{website}{Defaults to NULL. Website name, can be set once per session
with \code{\link[=cas_set_options]{cas_set_options()}}. This will be used as a second level folder and
may be used elsewhere to describe the dataset.}
}
\value{
Logical, TRUE if website folder exists, FALSE if it does not.
}
\description{
Parameters can be left to NULL; it will then rely on parameters set with \code{cas_set_options()}
}
