% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_convert_db_type.R
\name{cas_convert_db_type}
\alias{cas_convert_db_type}
\title{Convert database type, e.g. from DuckDB to SQLite}
\usage{
cas_convert_db_type(
  source_db_type,
  destination_db_type,
  disconnect_db = FALSE,
  ...
)
}
\arguments{
\item{source_db_type}{A database type, such as "DuckDB" or "SQLite". Must be declared explicitly.}

\item{destination_db_type}{A database type, such as "DuckDB" or "SQLite". Must be declared explicitly.}
}
\description{
Convert database type, e.g. from DuckDB to SQLite
}
