% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_count.R
\name{cas_count}
\alias{cas_count}
\title{Count strings in a corpus}
\usage{
cas_count(
  corpus,
  pattern,
  text = text,
  group_by = date,
  ignore_case = TRUE,
  drop_na = TRUE,
  fixed = FALSE,
  full_words_only = FALSE,
  pattern_column_name = pattern,
  n_column_name = n,
  locale = "en"
)
}
\arguments{
\item{corpus}{A textual corpus as a data frame.}

\item{pattern}{A character vector of one or more words or strings to be
counted.}

\item{text}{Defaults to \code{text}. The unquoted name of the column of the corpus
data frame to be used for matching.}

\item{group_by}{Defaults to NULL. If given, the unquoted name of the column
to be used for grouping (e.g. date, or doc_id, or source, etc.)}

\item{ignore_case}{Defaults to TRUE.}

\item{drop_na}{Defaults to TRUE. If TRUE, all rows where either \code{text} or
\code{group_by} column is NA are removed before further processing.}

\item{full_words_only}{Defaults to FALSE. If FALSE, string is counted even
when the it is found in the middle of a word (e.g. if FALSE, "ratio" would
be counted as match in the word "irrational").}

\item{pattern_column_name}{Defaults to \code{word}. The unquoted name of the
column to be used for the word in the output.}

\item{n_column_name}{Defaults to \code{n}. The unquoted name of the column to be
used for the count in the output.}

\item{locale}{Locale to be used when ignore_case is set to TRUE. Passed to
\code{stringr::str_to_lower}, defaults to "en".}
}
\value{
A data frame
}
\description{
Count strings in a corpus
}
\examples{
\dontrun{
cas_count(
  corpus = corpus,
  pattern = c("dogs", "cats", "horses"),
  text = text,
  group_by = date,
  n_column_name = n
)
}

}
