% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_count_relative.R
\name{cas_count_relative}
\alias{cas_count_relative}
\title{Count strings in a corpus relative to the number of words}
\usage{
cas_count_relative(
  corpus,
  string,
  text = text,
  group_by = date,
  ignore_case = TRUE,
  fixed = FALSE,
  full_words_only = FALSE,
  string_column_name = string,
  n_column_name = n,
  locale = "en"
)
}
\arguments{
\item{corpus}{A textual corpus as a data frame.}

\item{string}{A character vector of one or more words or strings to be counted.}

\item{text}{Defaults to text. The unquoted name of the column of the corpus data frame to be used for matching.}

\item{group_by}{Defaults to NULL. If given, the unquoted name of the column to be used for grouping (e.g. date, or doc_id, or source, etc.)}

\item{ignore_case}{Defaults to TRUE.}

\item{full_words_only}{Defaults to FALSE. If FALSE, string is counted even when the it is found in the middle of a word (e.g. if FALSE, "ratio" would be counted as match in the word "irrational").}

\item{string_column_name}{Defaults to 'word'. The unquoted name of the column to be used for the word in the output (if \code{include_string} is set to TRUE, as per default).}

\item{n_column_name}{Defaults to 'n'. The unquoted name of the column to be used for the count in the output.}

\item{locale}{Locale to be used when ignore_case is set to TRUE. Passed to \code{stringr::str_to_lower}, defaults to "en".}
}
\value{
A data frame
}
\description{
Count strings in a corpus relative to the number of words
}
\examples{

\dontrun{
cas_count_relative(
  corpus = corpus,
  string = c("dogs", "cats", "horses"),
  text = text,
  group_by = date,
  n_column_name = n
)
}

}
