% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_count.R
\name{cas_count_total_words}
\alias{cas_count_total_words}
\title{Count total words in a dataset}
\usage{
cas_count_total_words(
  corpus,
  string = "[\\\\w']+",
  text = text,
  group_by = date,
  ignore_case = TRUE,
  n_column_name = n,
  locale = "en"
)
}
\arguments{
\item{corpus}{A textual corpus as a data frame.}

\item{locale}{}
}
\description{
Count total words in a dataset
}
