% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_delete.R
\name{cas_delete}
\alias{cas_delete}
\title{Delete rows from selected database table}
\usage{
cas_delete(
  table,
  id = NULL,
  batch = NULL,
  index_group = NULL,
  ask = TRUE,
  db_folder = NULL,
  db_connection = NULL,
  ...
)
}
\arguments{
\item{table}{Name of the table from where rows should be deleted.}

\item{id}{Defaults to NULL. A vector of id. Rows with the given id will be removed from the database.}

\item{batch}{Defaults to NULL. A vector of batch identigiers. Rows with the given batch id will be removed from the database.}

\item{index_group}{Defaults to NULL. A vector of "index_group" names. Rows with the given "index_group" will be removed from the database.}

\item{ask}{Defaults to TRUE. If TRUE, it runs a query checking how many rows would be deleted, and actually deletes them only after confirming.}

\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\description{
Delete rows from selected database table
}
\examples{
\dontrun{
if (interactive) {
  cas_delete(table = "contents_data", id = id_to_delete)
}
}
}
