% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_delete_corpus.R
\name{cas_delete_corpus}
\alias{cas_delete_corpus}
\title{Delete previously stored corpora written with \code{cas_write_corpus()}.}
\usage{
cas_delete_corpus(
  keep = 1,
  ask = TRUE,
  file_format = "parquet",
  partition = "year",
  token = "full_text",
  corpus_folder = "corpus",
  path = NULL,
  ...
)
}
\arguments{
\item{keep}{Numeric, defaults to 1. Number of corpus files to keep. Only the most recent files are kept.}

\item{file_format}{Defaults to "parquet". Currently, other options are not
implemented.}

\item{partition}{Defaults to NULL. If NULL, the parquet file is not
partitioned. "year" is a common alternative: if set to "year", the parquet
file is partitioned by year. If a \code{year} column does not exist, it is
created based on the assumption that a \code{date} column exists and it is (or
can be coerced to) a vector of class \code{Date}.}

\item{token}{Defaults to "full_text", which does not tokenise the text
column. If different from \code{full_text}, it is passed to
\code{tidytext::unnest_tokens} (see its help for details). Accepted values
include "words", "sentences", and "paragraphs". See
\code{?tidytext::unnest_tokens()} for details.}

\item{path}{Defaults to NULL. If NULL, path is set to the
project/website/export/dataset/file_format folder.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\description{
Typically used for file maintainance, especially when datasets are routinely updated.
}
