% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_delete.R, R/cas_delete_from_db.R
\name{cas_delete_from_db}
\alias{cas_delete_from_db}
\title{Delete rows from selected database table}
\usage{
cas_delete_from_db(
  id,
  table,
  ask = TRUE,
  db_connection = NULL,
  disconnect_db = FALSE,
  ...
)

cas_delete_from_db(
  id,
  table,
  ask = TRUE,
  db_connection = NULL,
  disconnect_db = FALSE,
  ...
)
}
\arguments{
\item{id}{Defaults to NULL. A vector of id. Rows with the given id will be
removed from the database.}

\item{table}{Name of the table. See readme for details.}

\item{ask}{Defaults to TRUE. If TRUE, it runs a query checking how many rows
would be deleted, and actually deletes them only after confirming.}

\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{disconnect_db}{Defaults to \code{FALSE}. If \code{FALSE}, leaves the connection
to database open.}

\item{...}{Passed to \code{cas_get_db_file()}.}

\item{batch}{Defaults to NULL. A vector of batch identifiers. Rows with the
given batch id will be removed from the database.}

\item{index_group}{Defaults to NULL. A vector of "index_group" names. Rows
with the given "index_group" will be removed from the database.}

\item{df}{A data frame. Must correspond with the type of data expected for
each table.}

\item{overwrite}{Logical, defaults to \code{FALSE}. If \code{TRUE}, checks if matching data
are previously held in the table and overwrites them. This should be used
with caution, as it may overwrite completely the selected table.}
}
\value{
Nothing, used for its side effects.

If successful, returns invisibly the same data frame provided as
input and written to the database. Returns silently \code{NULL}, if nothing is
added, e.g. because \code{use_db} is set to \code{FALSE}.
}
\description{
Delete rows from selected database table

Generic function for removing lines from a database table
}
\examples{
\dontrun{
if (interactive) {
  cas_delete_from_db(table = "contents_data", id = id_to_delete)
}
}

cas_set_options(
  base_folder = fs::path(tempdir(), "R", "castarter_data"),
  project = "example_project",
  website = "example_website"
)
cas_enable_db()


urls_df <- cas_build_urls(
  url = "https://www.example.com/news/",
  start_page = 1,
  end_page = 10
)

cas_write_to_db(
  df = urls_df,
  table = "index_id"
)

cas_delete_from_db(id = 3:6, table = "index_id", ask = FALSE)
}
\seealso{
Other database functions:
\code{\link[=cas_check_db_folder]{cas_check_db_folder()}},
\code{\link[=cas_check_use_db]{cas_check_use_db()}},
\code{\link[=cas_connect_to_db]{cas_connect_to_db()}},
\code{\link[=cas_create_db_folder]{cas_create_db_folder()}},
\code{\link[=cas_disable_db]{cas_disable_db()}},
\code{\link[=cas_disconnect_from_db]{cas_disconnect_from_db()}},
\code{\link[=cas_enable_db]{cas_enable_db()}},
\code{\link[=cas_get_db_settings]{cas_get_db_settings()}},
\code{\link[=cas_read_from_db]{cas_read_from_db()}},
\code{\link[=cas_set_db]{cas_set_db()}},
\code{\link[=cas_set_db_folder]{cas_set_db_folder()}},
\code{\link[=cas_write_to_db]{cas_write_to_db()}}
}
\concept{database functions}
