% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_database.R
\name{cas_disable_cache}
\alias{cas_disable_cache}
\title{Disable caching for the current session}
\usage{
cas_disable_cache()
}
\value{
Nothing, used for its side effects.
}
\description{
Disable caching for the current session
}
\examples{
\donttest{
if (interactive()) {
  cas_disable_cache()
}
}
}
\seealso{
Other database functions: 
\code{\link{cas_check_cache_folder}()},
\code{\link{cas_check_cache}()},
\code{\link{cas_connect_to_cache}()},
\code{\link{cas_create_cache_folder}()},
\code{\link{cas_disconnect_from_cache}()},
\code{\link{cas_enable_cache}()},
\code{\link{cas_get_cache_db}()},
\code{\link{cas_set_cache_db}()},
\code{\link{cas_set_cache_folder}()}
}
\concept{database functions}
