% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_database.R
\name{cas_disconnect_from_cache}
\alias{cas_disconnect_from_cache}
\title{Ensure that connection to cache is disconnected consistently}
\usage{
cas_disconnect_from_cache(
  cache = NULL,
  cache_connection = NULL,
  disconnect_db = TRUE,
  language = castarter::cas_get_language()
)
}
\arguments{
\item{cache}{Defaults to NULL. If given, it should be given either TRUE or FALSE. Typically set with \code{cas_enable_cache()} or \code{cas_disable_cache()}.}

\item{cache_connection}{Defaults to NULL. If NULL, and caching is enabled, \code{castarter} will use a local sqlite database. A custom connection to other databases can be given (see vignette \code{caching} for details).}

\item{disconnect_db}{Defaults to TRUE. If FALSE, leaves the connection to cache open.}

\item{language}{Defaults to language set with \code{cas_set_language()}; if not set, "en". Use "all_available" to keep all languages. For available language values, see https://www.wikidata.org/wiki/Help:Wikimedia_language_codes/lists/all}
}
\value{
Nothing, used for its side effects.
}
\description{
Ensure that connection to cache is disconnected consistently
}
\examples{
if (interactive()) {
  cas_disconnect_from_cache()
}
}
\seealso{
Other database functions: 
\code{\link{cas_check_cache_folder}()},
\code{\link{cas_check_cache}()},
\code{\link{cas_connect_to_cache}()},
\code{\link{cas_create_cache_folder}()},
\code{\link{cas_disable_cache}()},
\code{\link{cas_enable_cache}()},
\code{\link{cas_get_cache_db}()},
\code{\link{cas_set_cache_db}()},
\code{\link{cas_set_cache_folder}()}
}
\concept{database functions}
