% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_download.R
\name{cas_download_index}
\alias{cas_download_index}
\title{Downloads index files systematically, and stores details about the download
in a local database}
\usage{
cas_download_index(
  download_df = NULL,
  index_group = NULL,
  file_format = "html",
  overwrite_file = FALSE,
  create_folder_if_missing = NULL,
  wait = 1,
  pause_base = 2,
  pause_cap = 256,
  pause_min = 4,
  sample = FALSE,
  retry_times = 8,
  terminate_on = 404,
  user_agent = NULL,
  download_again = FALSE,
  download_again_if_status_is_not = NULL,
  ...
)
}
\description{
Downloads index files systematically, and stores details about the download
in a local database
}
