% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_cache.R
\name{cas_enable_cache}
\alias{cas_enable_cache}
\title{Enable caching for the current session}
\usage{
cas_enable_cache(SQLite = TRUE)
}
\arguments{
\item{SQLite}{Logical, defaults to TRUE. Set to FALSE to use custom database options. See \code{cas_set_cache_db()} for details.}
}
\value{
Nothing, used for its side effects.
}
\description{
Enable caching for the current session
}
\examples{
\donttest{
if (interactive()) {
  cas_enable_cache()
}
}
}
