% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_explorer.R
\name{cas_explorer}
\alias{cas_explorer}
\title{Run the Shiny Application}
\usage{
cas_explorer(
  corpus = castarter::cas_demo_corpus,
  default_pattern = NULL,
  title = "castarter",
  collect = FALSE,
  advanced = FALSE,
  custom_head_html = "<meta name=\\"referrer\\" content=\\"no-referrer\\" />",
  footer_html = shiny::tagList(),
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  ...
)
}
\arguments{
\item{collect}{Defaults to FALSE. If TRUE, retrieves the corpus in memory,
even if is originally read from a parquet file or a database. With
\code{arrow} version before 14 that do not have full support of using stringr
in context, setting this to TRUE is probably advisable (currently,
depending on the arrow version, there may be issues where upper/lower case
is not ignored).}

\item{custom_head_html}{Chunk of code to be included in the app's \verb{<head>}.
This can be used, e.g., for custom analytics snippets. The default value,
\verb{<meta name="referrer" content="no-referrer" />} asks the browser not to
include the source website when following links to external websites.}

\item{onStart}{A function that will be called before the app is actually run.
This is only needed for \code{shinyAppObj}, since in the \code{shinyAppDir}
case, a \code{global.R} file can be used for this purpose.}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}

\item{enableBookmarking}{Can be one of \code{"url"}, \code{"server"}, or
\code{"disable"}. The default value, \code{NULL}, will respect the setting from
any previous calls to  \code{\link[shiny:enableBookmarking]{enableBookmarking()}}. See \code{\link[shiny:enableBookmarking]{enableBookmarking()}}
for more information on bookmarking your app.}

\item{uiPattern}{A regular expression that will be applied to each \code{GET}
request to determine whether the \code{ui} should be used to handle the
request. Note that the entire request path must match the regular
expression in order for the match to be considered successful.}

\item{...}{arguments to pass to golem_opts. See \code{?golem::get_golem_options}
for more details.}
}
\description{
Run the Shiny Application
}
