% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_export.R
\name{cas_export_tables}
\alias{cas_export_tables}
\title{Export database tables to another format such as csv}
\usage{
cas_export_tables(
  path = NULL,
  file_format = "csv.gz",
  tables = NULL,
  db_connection = NULL,
  disconnect_db = FALSE,
  db_folder = NULL,
  ...
)
}
\arguments{
\item{path}{Defaults to NULL. If NULL, path is set to the
project/website/export/file_format folder.}

\item{file_format}{Defaults to "csv.gz", i.e. compressed csv files. All
formats supported by \code{\link[readr:write_csv]{readr::write_csv()}} are valid.}

\item{tables}{Defaults to NULL. If NULL, all database tables are exported. If
given, names of the database tables to export. See
\code{vignette("castarter-database")} for details.}

\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{disconnect_db}{Defaults to \code{FALSE}. If \code{FALSE}, leaves the connection
to database open.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\description{
Export database tables to another format such as csv
}
\examples{
\dontrun{
if (interactive) {
  cas_export_tables(file_format = "csv")
}
}
}
