% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cass_utils.R
\name{cas_get_base_folder}
\alias{cas_get_base_folder}
\title{Get base folder under which files will be stored.}
\usage{
cas_get_base_folder(
  ...,
  level = c("base", "project", "website"),
  custom_path = NULL
)
}
\arguments{
\item{...}{Passed to \code{\link[=cas_get_options]{cas_get_options()}}.}

\item{level}{Defaults to "base". Valid values are "website", "project",
and "base".}

\item{custom_path}{Defaults to NULL. If given, all other parameters and
settings are ignored, and this value is returned.}
}
\value{
Path to the folder well all project files are stored.
}
\description{
The path is computed based on inputs and settings; it may or may not exist.
}
\examples{


cas_set_options(
  base_folder = fs::path(tempdir(), "R", "castarter_data"),
  db_folder = fs::path(tempdir(), "R", "castarter_data"),
  project = "example_project",
  website = "example_website"
)

cas_get_base_folder()

cas_get_base_folder(level = "website")
cas_get_base_folder(level = "project")
cas_get_base_folder(level = "base")

cas_get_base_folder(custom_path = fs::path(tempdir(), "custom_path"))
}
