% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_get_base_path.R
\name{cas_get_base_path}
\alias{cas_get_base_path}
\title{Build full path to base working folder}
\usage{
cas_get_base_path(
  create_folder_if_missing = NULL,
  custom_path = NULL,
  custom_folder = NULL,
  index = FALSE,
  file_format = "html",
  ...
)
}
\arguments{
\item{create_folder_if_missing}{Logical, defaults to NULL. If NULL, it will ask
before creating a new folder. If TRUE, it will create it without asking.}

\item{custom_path}{Defaults to NULL. If given, all other parameters and
settings are ignored, and this value is returned.}

\item{file_format}{}

\item{...}{Passed to \code{\link[=cas_get_options]{cas_get_options()}}.}
}
\value{
Path to base folder. A character vector of length one of class \code{fs_path}.
}
\description{
Build full path to base working folder
}
