% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_cache.R
\name{cas_get_cache_db}
\alias{cas_get_cache_db}
\title{Get database connection settings from the environment}
\usage{
cas_get_cache_db()
}
\value{
A list with all database parameters as stored in environment variables.
}
\description{
Typically set with \code{cas_set_cache_db()}
}
\examples{

cas_get_cache_db()
}
