% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_get_corpus_path.R
\name{cas_get_corpus_path}
\alias{cas_get_corpus_path}
\title{Get path to folder where the corpus is stored.}
\usage{
cas_get_corpus_path(
  ...,
  corpus_folder = "corpus",
  file_format = "parquet",
  partition = NULL,
  token = "full_text"
)
}
\arguments{
\item{...}{Passed to \code{cas_get_db_file()}.}

\item{file_format}{Defaults to "parquet". Currently, other options are not
implemented.}

\item{partition}{Defaults to \code{NULL}. If \code{NULL}, the parquet file is not
partitioned. "year" is a common alternative: if set to "year", the parquet
file is partitioned by year. If a \code{year} column does not exist, it is
created based on the assumption that a \code{date} column exists and it is (or
can be coerced to) a vector of class \code{Date}.}

\item{token}{Defaults to "full_text", which does not tokenise the text
column. If different from \code{full_text}, it is passed to
\code{\link[tidytext:unnest_tokens]{tidytext::unnest_tokens()}} (see its help for details). Accepted values
include "words", "sentences", and "paragraphs".}
}
\description{
Get path to folder where the corpus is stored.
}
\examples{
\dontrun{
 cas_get_corpus_path()
}
}
