% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_utils.R
\name{cas_get_db}
\alias{cas_get_db}
\title{Get connection to database with details about current website}
\usage{
cas_get_db(base_folder = NULL, project = NULL, website = NULL)
}
\arguments{
\item{base_folder}{Defaults to NULL, can be set once per session with \code{cas_set_options()}. A path to a location used for storing html and other project files. If the folder does not exist, it will be created. If not given, and not previously set as environment variable, defaults to \code{castarter_data}.}

\item{project}{Defaults to NULL. Project name, can be set once per session with \code{cas_set_options()}. This will be used as first level folder and may be used elsewhere to describe the dataset.}

\item{website}{Defaults to NULL. Website name, can be set once per session with \code{cas_set_options()}. This will be used as a second level folder and may be used elsewhere to describe the dataset.}
}
\value{

}
\description{
Get connection to database with details about current website
}
\examples{

cas_get_db(
  base_folder = fs::path_temp(),
  project = "example_project",
  website = "example_website"
)
}
