% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_download.R
\name{cas_get_files_to_download}
\alias{cas_get_files_to_download}
\title{Create a data frame with not yet downloaded files}
\usage{
cas_get_files_to_download(
  urls = NULL,
  index = FALSE,
  index_group = NULL,
  ignore_id = TRUE,
  desc_id = FALSE,
  batch = NULL,
  create_folder_if_missing = NULL,
  custom_folder = NULL,
  custom_path = NULL,
  file_format = "html",
  db_connection = NULL,
  download_again = FALSE,
  download_again_if_status_is_not = NULL,
  ...
)
}
\arguments{
\item{urls}{Defaults to \code{NULL}. If given, it should correspond with a data
frame with at least two columns named \code{id} and \code{url}. If not given, an
attempt will be made to load it from the local database.}

\item{index}{Logical, defaults to FALSE. If TRUE, downloaded files will be
considered \code{index} files. If not, they will be considered \code{contents} files.
See Readme for a more extensive explanation.}

\item{desc_id}{Logical, defaults to \code{FALSE}. If \code{TRUE}, results are returned with
highest id first.}

\item{batch}{An integer, defaults to \code{NULL}. If not given, a check is performed
in the database to find if previous downloads have taken place. If so, by
default, the current batch will be one unit higher than the highest batch
number found in the database.}

\item{file_format}{Defaults to \code{html}. Used for storing files in dedicated
folders, but also for determining processing options. For example, if a
sitemap is downloaded as an index with \code{file_format} set to xml, it will be
processed accordingly. If it is stored as xml.gz, it will be automatically
decompressed for correct processing.}

\item{download_again_if_status_is_not}{Defaults to \code{NULL}. If given, it must a
status code as integer, typically \code{200L}, or \code{c(200L, 404L)}.}

\item{...}{
  Arguments passed on to \code{\link{cas_get_base_folder}}
  \describe{
    \item{\code{custom_path}}{Defaults to NULL. If given, all other parameters and
settings are ignored, and this value is returned.}
  }}
}
\value{
A data frame with four columns: \code{id}, \code{url}, \code{path} and \code{type}
}
\description{
Create a data frame with not yet downloaded files
}
