% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_get_files_to_extract.R
\name{cas_get_files_to_extract}
\alias{cas_get_files_to_extract}
\title{Get path to (locally available) files to be extracted}
\usage{
cas_get_files_to_extract(
  id = NULL,
  ignore_id = TRUE,
  custom_path = NULL,
  index = FALSE,
  store_as_character = TRUE,
  check_previous = TRUE,
  db_connection = NULL,
  file_format = "html",
  sample = FALSE,
  keep_if_status = 200,
  only_available = TRUE,
  ...
)
}
\arguments{
\item{id}{Defaults to \code{NULL}, identifiers to process when extracting. If given,
must be a numeric vector, logically corresponding to the identifiers in the
\code{id} column, e.g. as returned by \code{\link[=cas_read_db_contents_id]{cas_read_db_contents_id()}}.}

\item{ignore_id}{Defaults to \code{TRUE}. If \code{TRUE}, it checks if identifiers have
been added to the local ignore list, typically with \code{\link[=cas_ignore_id]{cas_ignore_id()}}, and
as retrieved with \code{\link[=cas_read_db_ignore_id]{cas_read_db_ignore_id()}}. It can also be a numeric
vector of identifiers: the given identifiers will not be processed. If
\code{FALSE}, items will be processed normally.}

\item{index}{Logical, defaults to FALSE. If TRUE, downloaded files will be
considered \code{index} files. If not, they will be considered \code{contents} files.
See Readme for a more extensive explanation.}

\item{store_as_character}{Logical, defaults to \code{TRUE}. If \code{TRUE}, it converts to
character all extracted contents before writing them to database. This
reduces issues of type conversions with the default database backend (for
example, SQLite automatically converts dates to numeric) or using different
backends. This implies you will need to set data types when you read the
database, but it also means that you can consistently expect all columns to
be character vectors, which in one form or another are consistently
implemented across database backends. Set to \code{FALSE} if you want to remain in
control of column types.}

\item{check_previous}{Logical, defaults to \code{TRUE}. If \code{FALSE}, no check will be
conducted to verify if the same content had been previously extracted. If
FALSE, \code{write_to_db} must be set (or will be set) to \code{FALSE}, to prevent
duplication of data.}

\item{file_format}{Defaults to \code{html}. Used for storing files in dedicated
folders, but also for determining processing options. For example, if a
sitemap is downloaded as an index with \code{file_format} set to xml, it will be
processed accordingly. If it is stored as xml.gz, it will be automatically
decompressed for correct processing.}

\item{sample}{Defaults to FALSE. If TRUE, the download order is randomised.
If a numeric is given, the download order is randomised and at most the
given number of items is downloaded.}

\item{keep_if_status}{Defaults to 200. Keep only if recorded download status
matches the given status.}

\item{only_available}{Defaults to TRUE. If TRUE, returns only files available
locally. If FALSE, returns also path to files that according to logging
data have already been downloaded, yet are not available where expected.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\description{
Mostly used internally by \link{cas_extract} or for troubleshooting.
}
\examples{
#'
\dontrun{
if (interactive) {
  cas_get_files_to_extract()
}
}
}
