% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_download.R
\name{cas_get_urls_df}
\alias{cas_get_urls_df}
\title{Checks that a given input corresponds to the format expected of a download
data frame, consistently returns expected format}
\usage{
cas_get_urls_df(urls = NULL, index = FALSE, index_group = NULL, ...)
}
\arguments{
\item{url}{A character vector or a data frame with at least two columns, \code{id}
and \code{url}}
}
\value{
Consistently returns a data frame with at least two columns: a
numeric \code{id} column, and a character \code{url} column.
}
\description{
Checks that a given input corresponds to the format expected of a download
data frame, consistently returns expected format
}
\examples{

cas_get_urls_df(c(
  "https://example.com/a/",
  "https://example.com/b/"
))
}
