% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_ia_check.R
\name{cas_ia_check}
\alias{cas_ia_check}
\title{Gets an Archive.org Wayback Machine URL}
\usage{
cas_ia_check(
  url = NULL,
  wait = 1,
  retry_times = 3,
  pause_base = 2,
  pause_cap = 512,
  pause_min = 4,
  db_connection = NULL,
  disconnect_db = FALSE,
  check_db = TRUE,
  write_db = TRUE,
  output_only_newly_checked = FALSE,
  ...
)
}
\arguments{
\item{url}{A charachter vector of length one, a url.}

\item{wait}{Defaults to 1. Number of seconds to wait between downloading one
page and the next. Can be increased to reduce server load, or can be set to
0 when this is not an issue.}

\item{retry_times}{Defaults to 3. Number of times to retry download in case
of errors.}

\item{check_db}{Defaults to \code{TRUE}. If \code{TRUE}, checks if given URL has already
been checked in local database, and queries APIs only for URLs that have
not been previously checked.}

\item{write_db}{Defaults to \code{TRUE}. If \code{TRUE}, writes result to a local database.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\value{
A url linking to the version on the Internet Archive.
}
\description{
For details on API access to the Wayback Machine see:
https://archive.org/help/wayback_api.php
}
\details{
For an R package facilitating more extensive interaction with the API, see:
https://github.com/hrbrmstr/wayback

Integration with Wayback CDX Server API to be considered.
}
