% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_kwic.R
\name{cas_kwic}
\alias{cas_kwic}
\title{Adds a column with n words before and after the selected string to see keywords in context}
\usage{
cas_kwic(
  corpus,
  string,
  text = text,
  words_before = 5,
  words_after = 5,
  ignore_case = TRUE,
  full_words_only = FALSE
)
}
\arguments{
\item{corpus}{A textual corpus as a data frame.}

\item{string}{A string, typically of one or more words, to be used to break text. Should be of length 1 or length equal to the number of rows.}

\item{text}{Defaults to text. The unquoted name of the column of the corpus data frame to be used for matching.}

\item{words_before}{Integer, defaults to 5. Number of columns to include in the \code{before} column.}

\item{words_after}{Integer, defaults to 5. Number of columns to include in the \code{after} column.}

\item{ignore_case}{Defaults to TRUE.}

\item{full_words_only}{Defaults to FALSE. If FALSE, string is counted even when the it is found in the middle of a word (e.g. if FALSE, "ratio" would be counted as match in the word "irrational").}
}
\value{
A data frame (a tibble), with the same columns as input, plus three columns: before, string, and after. Only rows where the string is found are included.
}
\description{
Adds a column with n words before and after the selected string to see keywords in context
}
