% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_read_corpus.R
\name{cas_read_corpus}
\alias{cas_read_corpus}
\title{Read datasets created with \code{cas_write_dataset}}
\usage{
cas_read_corpus(
  ...,
  update = FALSE,
  path = NULL,
  file_format = "parquet",
  partition = NULL,
  token = "full_text",
  corpus_folder = "corpus"
)
}
\arguments{
\item{...}{Passed to \code{cas_get_db_file()}.}

\item{update}{Logical, defaults to FALSE. If FALSE, just checks if relevant
corpus has been previously stored. If TRUE, it checks if more recent
contents are available in the local database.}

\item{path}{Defaults to NULL. If NULL, path is set to the
project/website/export/dataset/file_format folder.}

\item{file_format}{Defaults to "parquet". Currently, other options are not
implemented.}

\item{partition}{Defaults to NULL. If NULL, the parquet file is not
partitioned. "year" is a common alternative: if set to "year", the parquet
file is partitioned by year. If a \code{year} column does not exist, it is
created based on the assumption that a \code{date} column exists and it is (or
can be coerced to) a vector of class \code{Date}.}

\item{token}{Defaults to "full_text", which does not tokenise the text
column. If different from \code{full_text}, it is passed to
\code{tidytext::unnest_tokens} (see its help for details). Accepted values
include "words", "sentences", and "paragraphs". See
\code{?tidytext::unnest_tokens()} for details.}
}
\value{
A dataset as \code{ArrowObject}
}
\description{
Read datasets created with \code{cas_write_dataset}
}
\examples{
\dontrun{
cas_read_corpus()
}
}
