% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_db_contents.R
\name{cas_read_db_contents}
\alias{cas_read_db_contents}
\title{Read contents from local database}
\usage{
cas_read_db_contents(
  use_db = NULL,
  db_connection = NULL,
  disconnect_db = TRUE,
  db_folder = NULL,
  project = cas_get_options()$project,
  website = cas_get_options()$website
)
}
\arguments{
\item{use_db}{Defaults to NULL. If given, it should be given either TRUE or
FALSE. Typically set with \code{cas_enable_db()} or \code{cas_disable_db()}.}

\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{disconnect_db}{Defaults to TRUE. If FALSE, leaves the connection to database open.}
}
\value{
A data frame with three columns and data stored in the \code{contents_id}
table of the local database. The data frame has zero rows if the database
does not exist or no data was previously stored there.
}
\description{
Read contents from local database
}
\examples{
cas_set_options(
  base_folder = fs::path(tempdir(), "R", "castarter_data"),
  db_folder = fs::path(tempdir(), "R", "castarter_data"),
  project = "example_project",
  website = "example_website"
)
cas_enable_db()


urls_df <- cas_build_urls(
  url_beginning = "https://www.example.com/news/",
  start_page = 1,
  end_page = 10
)

cas_write_db_contents(urls = urls_df)

cas_read_db_contents()
}
