% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_read_db_contents_data.R
\name{cas_read_db_contents_data}
\alias{cas_read_db_contents_data}
\title{Read contents data from local database}
\usage{
cas_read_db_contents_data(db_connection = NULL, db_folder = NULL, ...)
}
\arguments{
\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\description{
Read contents data from local database
}
