% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_db_index.R
\name{cas_read_db_download}
\alias{cas_read_db_download}
\title{Read index from local database}
\usage{
cas_read_db_download(
  index = FALSE,
  db_connection = NULL,
  db_folder = NULL,
  ...
)
}
\arguments{
\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\value{
A data frame with three columns and data stored in the \code{index_id}
table of the local database. The data frame has zero rows if the database
does not exist or no data was previously stored there.
}
\description{
Read index from local database
}
\examples{
cas_set_options(
  base_folder = fs::path(tempdir(), "R", "castarter_data"),
  db_folder = fs::path(tempdir(), "R", "castarter_data"),
  project = "example_project",
  website = "example_website"
)
cas_enable_db()


urls_df <- cas_build_urls(
  url_beginning = "https://www.example.com/news/",
  start_page = 1,
  end_page = 10
)

cas_write_db_index(urls = urls_df)

cas_read_db_index()
}
