% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_ia_check.R
\name{cas_read_db_ia}
\alias{cas_read_db_ia}
\title{Read status on the Internet Archive of given URLs}
\usage{
cas_read_db_ia(db_connection = NULL, db_folder = NULL, ...)
}
\arguments{
\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\description{
Read status on the Internet Archive of given URLs
}
