% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_ignore_id.R
\name{cas_read_db_ignore_id}
\alias{cas_read_db_ignore_id}
\title{Read identifiers to be ignored from the local database}
\usage{
cas_read_db_ignore_id(
  db_connection = NULL,
  db_folder = NULL,
  index_group = NULL,
  disconnect_db = TRUE,
  ...
)
}
\value{
A data frame with a single column, \code{id}
}
\description{
Read identifiers to be ignored from the local database
}
\examples{
cas_set_options(
  base_folder = fs::path(tempdir(), "R", "cas_read_db_ignore_id"),
  db_folder = fs::path(tempdir(), "R", "cas_read_db_ignore_id"),
  project = "example_project",
  website = "example_website"
)
cas_enable_db()


cas_write_db_ignore_id(id = sample(x = 1:100, size = 10))

cas_read_db_ignore_id()
}
