% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_get_sitemap.R
\name{cas_read_db_sitemap}
\alias{cas_read_db_sitemap}
\title{Read sitemap from local database}
\usage{
cas_read_db_sitemap(db_connection = NULL, db_folder = NULL, ...)
}
\arguments{
\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\value{
A data frame with three columns and data stored in the \code{sitemap}
table of the local database. The data frame has zero rows if the database
does not exist or no data was previously stored there.
}
\description{
Read sitemap from local database
}
\examples{
cas_set_options(
  base_folder = fs::path(fs::path_temp(), "R", "castarter_data"),
  db_folder = fs::path(fs::path_temp(), "R", "castarter_data"),
  project = "example_project",
  website = "example_website"
)
cas_enable_db()

cas_write_db_sitemap(sitemap = "https://example.com/sitemap.xml")

cas_read_db_sitemap()
}
