% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_write_db_urls.R
\name{cas_read_db_urls}
\alias{cas_read_db_urls}
\title{Read urls stored in the local database}
\usage{
cas_read_db_urls(
  index = FALSE,
  db_connection = NULL,
  db_folder = NULL,
  index_group = NULL,
  ...
)
}
\arguments{
\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\description{
Read urls stored in the local database
}
