% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_index.R
\name{cas_read_index}
\alias{cas_read_index}
\title{Read index from local database}
\usage{
cas_read_index(use_db = NULL, db_connection = NULL, disconnect_db = TRUE)
}
\arguments{
\item{use_db}{Defaults to NULL. If given, it should be given either TRUE or FALSE. Typically set with \code{cas_enable_db()} or \code{cas_disable_db()}.}

\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database. If given, must be a connection object or a list with relevant connection settings (see example).}

\item{disconnect_db}{Defaults to TRUE. If FALSE, leaves the connection to database open.}
}
\description{
Read index from local database
}
