% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_reset.R
\name{cas_reset_db_ignore_id}
\alias{cas_reset_db_ignore_id}
\title{Removes from the local database all identifiers included in the ignore list}
\usage{
cas_reset_db_ignore_id(db_connection = NULL, db_folder = NULL, ask = TRUE, ...)
}
\arguments{
\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{ask}{Logical, defaults to TRUE. If set to FALSE, the relevant table
will be deleted without asking for confirmation from the user.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\description{
Removes from the local database all identifiers included in the ignore list
}
\examples{
cas_set_options(
  base_folder = fs::path(tempdir(), "R", "cas_reset_db_ignore_id"),
  db_folder = fs::path(tempdir(), "R", "cas_reset_db_ignore_id"),
  project = "example_project",
  website = "example_website"
)
cas_enable_db()


cas_write_db_ignore_id(id = sample(x = 1:100, size = 10))

cas_read_db_ignore_id()

cas_reset_db_ignore_id(ask = FALSE)

cas_read_db_ignore_id()
}
