% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_reset.R
\name{cas_reset_db_index_id}
\alias{cas_reset_db_index_id}
\title{Removes from the local database the folder where links to index associated
with their id are stored}
\usage{
cas_reset_db_index_id(db_connection = NULL, db_folder = NULL, ask = TRUE, ...)
}
\arguments{
\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{ask}{Logical, defaults to TRUE. If set to FALSE, the relevant table
will be deleted without asking for confirmation from the user.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\description{
Removes from the local database the folder where links to index associated
with their id are stored
}
