% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_reset.R
\name{cas_reset_download_contents}
\alias{cas_reset_download_contents}
\title{Delete all files and database records for the contents pages of the current
website}
\usage{
cas_reset_download_contents(
  batch = NULL,
  file_format = "html",
  db_connection = NULL,
  db_folder = NULL,
  ask = TRUE,
  ...
)
}
\arguments{
\item{batch}{Defaults to NULL. If given, only files and records related to
the given batch are removed. If not given, all contents files are removed.}

\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{ask}{Logical, defaults to TRUE. If set to FALSE, the relevant table
will be deleted without asking for confirmation from the user.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\description{
Delete all files and database records for the contents pages of the current
website
}
