% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_restore.R
\name{cas_restore}
\alias{cas_restore}
\title{Restore files from compressed files}
\usage{
cas_restore(
  restore_to = NULL,
  restore_from = NULL,
  file_format = "tar.gz",
  index = FALSE,
  contents = FALSE,
  batch = NULL,
  db_connection = NULL,
  db_folder = NULL,
  ...
)
}
\arguments{
\item{restore_to}{Path to archive directory, defaults to NULL. If NULL, path
is set to the project/website/archive folder.}

\item{restore_from}{Path to archive directory, defaults to NULL. If NULL,
path is set to the project/website/archive folder.}

\item{file_format}{Defaults to "tar.gz", to ensure cross-platform
compatibility. No other formats are supported at this stage.}

\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\value{
A path to the base folder where files are stored. Corresponds to
\code{restore_to} if given, or to a temporary folder if \code{restore_to} is set to
NULL.
}
\description{
Restore files from compressed files
}
