% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_database.R
\name{cas_set_cache_db}
\alias{cas_set_cache_db}
\title{Set database connection settings for the session}
\usage{
cas_set_cache_db(
  db_settings = NULL,
  driver = NULL,
  host = NULL,
  port,
  database,
  user,
  pwd
)
}
\arguments{
\item{db_settings}{A list of database connection settings (see example)}

\item{driver}{A database driver. Common database drivers include \code{MySQL}, \code{PostgreSQL}, and \code{MariaDB}. See \code{unique(odbc::odbcListDrivers()[[1]])} for a list of locally available drivers.}

\item{host}{Host address, e.g. "localhost".}

\item{port}{Port to use to connect to the database.}

\item{database}{Database name.}

\item{user}{Database user name.}

\item{pwd}{Password for the database user.}
}
\value{
A list with all given parameters (invisibly).
}
\description{
Set database connection settings for the session
}
\examples{
\donttest{
if (interactive()) {

  # Settings can be provided either as a list
  db_settings <- list(
    driver = "MySQL",
    host = "localhost",
    port = 3306,
    database = "castarter",
    user = "secret_username",
    pwd = "secret_password"
  )

  cas_set_cache_db(db_settings)

  # or as parameters

  cas_set_cache_db(
    driver = "MySQL",
    host = "localhost",
    port = 3306,
    database = "castarter",
    user = "secret_username",
    pwd = "secret_password"
  )
}
}
}
\seealso{
Other database functions: 
\code{\link{cas_check_cache_folder}()},
\code{\link{cas_check_cache}()},
\code{\link{cas_connect_to_cache}()},
\code{\link{cas_create_cache_folder}()},
\code{\link{cas_disable_cache}()},
\code{\link{cas_disconnect_from_cache}()},
\code{\link{cas_enable_cache}()},
\code{\link{cas_get_cache_db}()},
\code{\link{cas_set_cache_folder}()}
}
\concept{database functions}
