% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_cache.R
\name{cas_set_cache_folder}
\alias{cas_set_cache_folder}
\alias{cas_get_cache_folder}
\title{Set folder for caching data}
\usage{
cas_set_cache_folder(path = NULL)

cas_get_cache_folder(path = NULL)
}
\arguments{
\item{path}{A path to a location used for caching data. If the folder does not exist, it will be created.}
}
\value{
The path to the caching folder, if previously set; the same path as given to the function; or the default, \code{cas_data} is none is given.
}
\description{
Consider using a folder out of your current project directory, e.g. \code{cas_set_cache_folder("~/R/cas_data/")}: you will be able to use the same cache in different projects, and prevent cached files from being sync-ed if you use services such as Nextcloud or Dropbox.
}
\examples{
\donttest{
if (interactive()) {
  cas_set_cache_folder(fs::path(fs::path_home_r(), "R", "cas_data"))
}
}
cas_get_cache_folder()
}
\seealso{
Other caching functions: 
\code{\link{cas_check_cache_folder}()},
\code{\link{cas_check_cache}()},
\code{\link{cas_connect_to_cache}()},
\code{\link{cas_create_cache_folder}()},
\code{\link{cas_disable_cache}()},
\code{\link{cas_disconnect_from_cache}()},
\code{\link{cas_enable_cache}()},
\code{\link{cas_get_cache_db}()},
\code{\link{cas_set_cache_db}()}
}
\concept{caching functions}
