% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_options.R
\name{cas_set_options}
\alias{cas_set_options}
\title{Set key project parameters that determine the folder used for storing project files}
\usage{
cas_set_options(
  project = NULL,
  website = NULL,
  use_db = TRUE,
  base_folder = NULL,
  db_type = "SQLite",
  db_folder = NULL
)
}
\arguments{
\item{project}{Defaults to NULL. Project name, can be set once per session
with \code{\link[=cas_set_options]{cas_set_options()}}. This will be used as first level folder and may
be used elsewhere to describe the dataset.}

\item{website}{Defaults to NULL. Website name, can be set once per session
with \code{\link[=cas_set_options]{cas_set_options()}}. This will be used as a second level folder and
may be used elsewhere to describe the dataset.}

\item{use_db}{Defaults to TRUE. If TRUE, stores information about the
download process and extracted text in a local database.}

\item{base_folder}{Defaults to NULL, can be set once per session with
\code{\link[=cas_set_options]{cas_set_options()}}. A path to a location used for storing html and other
project files. If the folder does not exist, it will be created. If not
given, and not previously set as environment variable, defaults to
\code{castarter_data}.}

\item{db_folder}{Defaults to NULL. can be set once per session with
\code{\link[=cas_set_options]{cas_set_options()}} or \code{\link[=cas_set_db_folder]{cas_set_db_folder()}}. A path to a location used for
storing the database. If not given, and not previously set as environment
variable, defaults to \code{castarter_data}.}
}
\value{
Nothing, used for its side effects (setting options).
}
\description{
Your project folder can be anywhere on your file system. Considering that
this is where possibly a very large number of html files will be downloaded,
it is usually preferable to choose a location that is not included in live
backups. These settings determine the names given to these hierarchical
folders: \code{website} folder will be under \code{project} folder which will be under
the \code{base_folder}.
}
\examples{
cas_set_options(base_folder = fs::path(fs::path_temp(), "castarter_data"))
cas_options_list <- cas_get_options()
cas_options_list
}
\seealso{
Other settings:
\code{\link[=cas_get_options]{cas_get_options()}}
}
\concept{settings}
