% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_show_barchart.R
\name{cas_show_barchart_ggiraph}
\alias{cas_show_barchart_ggiraph}
\title{Creates interacative barchart with ggiraph}
\usage{
cas_show_barchart_ggiraph(ggobj, data_id = NULL, tooltip = NULL)
}
\arguments{
\item{ggobj}{A ggplot2 object, typically generated with \code{cas_show_gg_base()}}

\item{data_id}{Defaults to NULL. If given, unquoted name of column, passed to ggiraph.}

\item{tooltip}{Defaults to NULL. If given, unquoted name of column, passed to ggiraph.}
}
\value{
A girage/htmlwidget object
}
\description{
For detail on parameters, see https://davidgohel.github.io/ggiraph/articles/offcran/using_ggiraph.html
}
