% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_show_barchart.R
\name{cas_show_barchart_ggplot2}
\alias{cas_show_barchart_ggplot2}
\title{Creates barchart with ggplot2}
\usage{
cas_show_barchart_ggplot2(ggobj, position = "stack")
}
\arguments{
\item{ggobj}{A ggplot2 object, typically generated with \code{cas_show_gg_base()}}

\item{position}{Defaults to "stack". Available values include "dodge".}
}
\value{
A ggplot2 object
}
\description{
Creates barchart with ggplot2
}
