% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_summarise.R
\name{cas_summarise}
\alias{cas_summarise}
\title{Summarise for a given time period word counts, typically calculatd with \code{cas_count()}}
\usage{
cas_summarise(
  count_df,
  date = date,
  n = n,
  period = NULL,
  f = mean,
  every = 1L,
  before = 0L,
  after = 0L,
  complete = FALSE,
  auto_convert = FALSE
)
}
\arguments{
\item{count_df}{A data frame. Must include at least a column with a date or date-time column and a column with number of occurrences for the given time.}

\item{date}{Defaults to \code{date}. Unquoted name of a column having either date or date-time as class.}

\item{n}{Unquoted to \code{n}.  Unquoted name of a column having number of occurrences per time unit.}

\item{period}{Defaults to NULL. A string describing the time unit to be used for summarising. Possible values include "year", "quarter", "month", "day", "hour", "minute", "second", "millisecond".}

\item{f}{Defaults to \code{mean}. Function to be applied over n for all the values in a given time period. Common alterantives would be \code{mean} or \code{median}.}

\item{every}{Check \code{?slider::slide_period()} for details.}

\item{before}{Check \code{?slider::slide_period()} for details.}

\item{after}{Check \code{?slider::slide_period()} for details.}

\item{complete}{Check \code{?slider::slide_period()} for details.}

\item{auto_convert}{Defaults to FALSE. If FALSE, the date column is returned using the same format as the input; the minimun vale in the given group is used for reference (e.g. all values for January 2022 are summarised as 2021-01-01 it the data were originally given as dates.). If TRUE, it tries to adapt the output to the most intuitive correspondent type; for year, a numeric column with only the year number, for quarter in the format 2022.1, for month in the format 2022-01.}
}
\value{
A data frame with two columns: the name of the period, and the same name originally used for \code{n}.
}
\description{
Summarise for a given time period word counts, typically calculatd with \code{cas_count()}
}
\examples{
\dontrun{
# this assumes dates are provided in a column called date
corpus_df \%>\%
 cas_count(words = "example",
           group_by = date) \%>\%
 cas_summarise(period = "year")
}

}
