% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_write_corpus.R
\name{cas_write_corpus}
\alias{cas_write_corpus}
\title{Export the textual dataset for the current website}
\usage{
cas_write_corpus(
  path = NULL,
  file_format = "parquet",
  db_connection = NULL,
  db_folder = NULL,
  ...
)
}
\arguments{
\item{path}{Defaults to NULL. If NULL, path is set to the project/website/export/dataset/file_format folder.}

\item{file_format}{Defaults to "parquet". Currently, other options are not implemented.}

\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\description{
Export the textual dataset for the current website
}
