% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_write_db_contents_data.R
\name{cas_write_db_contents_data}
\alias{cas_write_db_contents_data}
\title{Write extracted contents to local database}
\usage{
cas_write_db_contents_data(
  contents_df,
  overwrite = FALSE,
  db_connection = NULL,
  disconnect_db = FALSE,
  quiet = FALSE,
  check_previous = TRUE,
  ...
)
}
\arguments{
\item{overwrite}{Logical, defaults to \code{FALSE}. If \code{TRUE}, checks if matching data
are previously held in the table and overwrites them. This should be used
with caution, as it may overwrite completely the selected table.}

\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{disconnect_db}{Defaults to \code{FALSE}. If \code{FALSE}, leaves the connection
to database open.}

\item{quiet}{Defaults to FALSE. If set to TRUE, messages on number of lines
added are not shown.}

\item{check_previous}{Defaults to TRUE. If set to FALSE, the given input is
stored in the database without checking if the same id had already been
stored.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\value{
Invisibly returns only new rows added.
}
\description{
If some IDs are already present in the database, only the new ones are
appended: IDs are expected to be unique.
}
\details{
Check for consistency in database columns: if new columns do not match
previous columns, it throws an error.
}
