% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_db_contents_id.R
\name{cas_write_db_contents_id}
\alias{cas_write_db_contents_id}
\title{Write contents URLs to local database}
\usage{
cas_write_db_contents_id(
  urls,
  overwrite = FALSE,
  db_connection = NULL,
  disconnect_db = FALSE,
  quiet = FALSE,
  check_previous = TRUE,
  ...
)
}
\arguments{
\item{urls}{A data frame with five columns, such as
\code{casdb_empty_contents_id}, or a character vector.}

\item{overwrite}{Logical, defaults to \code{FALSE}. If \code{TRUE}, checks if matching data
are previously held in the table and overwrites them. This should be used
with caution, as it may overwrite completely the selected table.}

\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{disconnect_db}{Defaults to \code{FALSE}. If \code{FALSE}, leaves the connection
to database open.}

\item{quiet}{Defaults to FALSE. If set to TRUE, messages on number of lines
added are not shown.}

\item{check_previous}{Defaults to TRUE. If set to FALSE, the given input is
stored in the database without checking if the same url had already been
stored.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\value{
Invisibly returns only new rows added.
}
\description{
If some URLs are already included in the database, it appends only the new
ones: URLs are expected to be unique.
}
\examples{

cas_set_options(
  base_folder = fs::path(tempdir(), "R", "castarter_data"),
  db_folder = fs::path(tempdir(), "R", "castarter_data"),
  project = "example_project",
  website = "example_website"
)
cas_enable_db()


urls_df <- cas_build_urls(
  url = "https://www.example.com/news/",
  start_page = 1,
  end_page = 10
)

cas_write_db_contents_id(urls = urls_df$url)

cas_read_db_contents_id()
}
