% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_ignore_id.R
\name{cas_write_db_ignore_id}
\alias{cas_write_db_ignore_id}
\alias{cas_ignore_id}
\title{Ignore a set of ids from the download or processing step}
\usage{
cas_write_db_ignore_id(
  id,
  db_folder = NULL,
  db_connection = NULL,
  disconnect_db = FALSE,
  ...
)

cas_ignore_id(
  id,
  db_folder = NULL,
  db_connection = NULL,
  disconnect_db = FALSE,
  ...
)
}
\arguments{
\item{id}{Defaults to NULL. A vector of id. Rows with the given id will be
added to the ignore table.}
}
\description{
There are two main use cases for this function:
\itemize{
\item a number of the files downloaded turned out to be irrelevant. Rather than delete any trace about them, it may be preferrable to just ignore them, so they are not processed when extracting data.
\item urls originally included for download, but not yet downloaded, should be ignored and not downloaded. This may or may not be a temporary arrangement, but it is considered useful to keep the urls in the database.
}
}
