% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_get_sitemap.R
\name{cas_write_db_sitemap}
\alias{cas_write_db_sitemap}
\title{Write index URLs to local database}
\usage{
cas_write_db_sitemap(
  sitemap,
  overwrite = FALSE,
  db_connection = NULL,
  disconnect_db = FALSE,
  ...
)
}
\arguments{
\item{sitemap}{A sitemap url, or a dataframe with at least one column named
\code{sitemap_url}.}

\item{overwrite}{Logical, defaults to \code{FALSE}. If \code{TRUE}, checks if matching data
are previously held in the table and overwrites them. This should be used
with caution, as it may overwrite completely the selected table.}

\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{disconnect_db}{Defaults to \code{FALSE}. If \code{FALSE}, leaves the connection
to database open.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\value{
Invisibly returns a data frame with a single column, \code{sitemap_url}.
}
\description{
If some URLs are already included in the database, it appends only the new
ones: URLs are expected to be unique.
}
\examples{
cas_set_options(
  base_folder = fs::path(fs::path_temp(), "R", "castarter_data"),
  db_folder = fs::path(fs::path_temp(), "R", "castarter_data"),
  project = "example_project",
  website = "example_website"
)
cas_enable_db()

cas_write_db_sitemap(sitemap = "https://example.com/sitemap.xml")

cas_read_db_sitemap()
}
