% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_index.R
\name{cas_write_index}
\alias{cas_write_index}
\title{Write index URLs to local database}
\usage{
cas_write_index(
  urls,
  use_db = NULL,
  overwrite = FALSE,
  db_connection = NULL,
  disconnect_db = TRUE
)
}
\arguments{
\item{urls}{A data frame with three columns, such as \code{casdb_empty_index_id}, or a character vector.}

\item{use_db}{Defaults to NULL. If given, it should be given either TRUE or
FALSE. Typically set with \code{cas_enable_db()} or \code{cas_disable_db()}.}

\item{overwrite}{Logical, defaults to FALSE. If TRUE, checks if matching data are previously held in the table and overwrites them. This should be used with caution, as it may overwrite completely the selected table.}

\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{disconnect_db}{Defaults to TRUE. If FALSE, leaves the connection to database open.}
}
\description{
If some URLs are already included in the database, it appends only the new ones: URLs are expected to be unique.
}
\examples{

cas_set_options(
base_folder = fs::path(tempdir(), "R", "castarter_data"),
db_folder = fs::path(tempdir(), "R", "castarter_data"),
project = "example_project",
website = "example_website"
)
cas_enable_db()


urls_df <- cas_build_urls(
  url_beginning = "https://www.example.com/news/",
  start_page = 1,
  end_page = 10
)

cas_write_index(urls = urls_df)

cas_read_index()
}
