% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{casdb_empty_index_id}
\alias{casdb_empty_index_id}
\title{Empty data frame with the same format as data stored in the \code{index_id} table}
\format{
A data frame with 0 rows and 3 columns:
\describe{
\item{id}{Numeric. Column meant for unique integer identifier corresponding to a unique url}
\item{url}{Character. A url.}
\item{type}{Character. A textual string, by default \code{index}.}
}
}
\usage{
casdb_empty_index_id
}
\description{
Empty data frame with the same format as data stored in the \code{index_id} table
}
\keyword{datasets}
