% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cass_build_urls.R, R/mod_casm_build_urls.R
\name{cass_build_urls}
\alias{cass_build_urls}
\title{Helps you define the parameters you need for building index urls}
\usage{
cass_build_urls()

cass_build_urls()
}
\value{
Nothing, but prints to the console the function call as created in the Shiny app.

Nothing, called for interactive use.
}
\description{
Helps you define the parameters you need for building index urls

Helps you define the parameters you need for building index urls
}
\examples{
\dontrun{
if (interactive) {
  cass_build_urls()
}
}
\dontrun{
if (interactive) {
  cass_build_urls()
}
}
}
