% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cass_download_csv_button.R
\name{cass_download_csv_app}
\alias{cass_download_csv_app}
\title{A minimal shiny app that demonstrates the functioning of related modules}
\usage{
cass_download_csv_app(df, type)
}
\arguments{
\item{df}{A data frame to be exported as csv.}
}
\value{
A shiny app
}
\description{
A minimal shiny app that demonstrates the functioning of related modules
}
\examples{

count_df <- castarter2::cas_count(
  corpus = castarter2::cas_demo_corpus,
  string = c("russia", "moscow")
) \%>\%
  cas_summarise(before = 15, after = 15)

# cass_cass_download_csv_app(count_df)
}
