% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_download.R
\name{cass_download_httr}
\alias{cass_download_httr}
\title{Downloads one file at a time with httr}
\usage{
cass_download_httr(
  download_df = NULL,
  index = FALSE,
  overwrite_file = FALSE,
  wait = 1,
  db_connection = NULL,
  random = FALSE,
  file_format = "html",
  retry_times = 3,
  download_again_if_status_is_not = NULL,
  ...
)
}
\arguments{
\item{download_df}{A data frame with four columns: \code{id}, \code{url}, \code{path}, \code{type}.}

\item{index}{Logical, defaults to FALSE. If TRUE, downloaded files will be
considered \code{index} files. If not, they will be considered \code{contents} files.
See Readme for a more extensive explanation.}

\item{overwrite_file}{Logical, defaults to FALSE.}

\item{wait}{Defaults to 1. Number of seconds to wait between downloading one
page and the next. Can be increased to reduce server load, or can be set to
0 when this is not an issue.}

\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{random}{Defaults to FALSE. If TRUE, the download order is randomised.
If a numeric is given, the download order is randomised and at most the
given number of items is downloaded.}

\item{retry_times}{Defaults to 3. Number of times to retry download in case
of errors.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\value{
Invisibly returns the full \code{httr} response.
}
\description{
Mostly used internally by \code{cas_download}.
}
