% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cass_get_base_path.R
\name{cass_get_base_path}
\alias{cass_get_base_path}
\title{Build full path to base working folder}
\usage{
cass_get_base_path(
  custom_path = NULL,
  custom_folder = NULL,
  index = FALSE,
  file_format = "html"
)
}
\arguments{
\item{custom_path}{Defaults to NULL. If given, all other parameters and settings are ignored, and folder is set to this value.}

\item{custom_folder}{}

\item{file_format}{}
}
\value{
Path to base folder. A character vector of length one of class \code{fs_path}.
}
\description{
Build full path to base working folder
}
