% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cass_utils.R
\name{cass_split}
\alias{cass_split}
\title{Split string into multiple inputs}
\usage{
cass_split(string, squish = TRUE, to_lower = TRUE)
}
\arguments{
\item{string}{A text string, typically a user input in a shiny app.}
}
\value{
A character vector
}
\description{
Split string into multiple inputs
}
\examples{
if (interactive()) {
cass_split("dogs, cats, horses")
}
}
