% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cass_utils.R
\name{cass_split_string}
\alias{cass_split_string}
\title{Split string into multiple inputs}
\usage{
cass_split_string(string, squish = TRUE, to_lower = TRUE, to_regex = FALSE)
}
\arguments{
\item{string}{A text string, typically a user input in a shiny app.}

\item{to_regex}{Defaults to FALSE. If TRUE collapses the split string,
separating each element with \code{|}.}
}
\value{
A character vector
}
\description{
Split string into multiple inputs
}
\examples{
cass_split_string("dogs, cats, horses")
cass_split_string(string = "dogs, cats, horses", to_regex = TRUE)
}
