% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_browse.R
\name{cas_browse}
\alias{cas_browse}
\title{Open in a browser a URL}
\usage{
cas_browse(
  index = FALSE,
  remote = FALSE,
  id = NULL,
  batch = NULL,
  index_group = NULL,
  file_format = "html",
  random = 1,
  ...
)
}
\arguments{
\item{index}{Logical, defaults to FALSE. If TRUE, downloaded files will be
considered \code{index} files. If not, they will be considered \code{contents} files.
See Readme for a more extensive explanation.}

\item{random}{Defaults to 1.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\description{
Open in a browser a URL
}
