% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_browse.R
\name{cas_browse}
\alias{cas_browse}
\title{Open in a browser a URL stored in the local database}
\usage{
cas_browse(
  index = FALSE,
  remote = TRUE,
  id = NULL,
  batch = NULL,
  index_group = NULL,
  file_format = "html",
  sample = 1,
  disconnect_db = TRUE,
  ...
)
}
\arguments{
\item{index}{Logical, defaults to FALSE. If TRUE, downloaded files will be
considered \code{index} files. If not, they will be considered \code{contents} files.
See Readme for a more extensive explanation.}

\item{remote}{Defaults to TRUE. If TRUE, opens relevant url online. If FALSE,
it opens the locally stored file.}

\item{sample}{Defaults to 1. By default, it opens one random url.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\description{
This function is typically used to check a web page when extracting links
from index, or contents from contents pages.
}
