% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_build_urls.R
\name{cas_build_urls}
\alias{cas_build_urls}
\title{URL builder}
\usage{
cas_build_urls(
  url_beginning,
  url_ending = "",
  start_page = 1,
  end_page = 10,
  increase_by = 1,
  date_format = "Ymd",
  start_date = NULL,
  end_date = Sys.Date(),
  date_separator = NULL,
  increase_date_by = "day",
  reversed_order = FALSE,
  type = "index"
)
}
\arguments{
\item{url_beginning}{First part of index link that does not change in other
index pages.}

\item{url_ending}{Part of index link appneded after the part of the link that
varies. If not relevant, may be left empty.}

\item{start_page}{If the urls include a numerical component, define first
number of the sequence. start_page defaults to 1.}

\item{end_page}{If the urls include a numerical component, define first
number of the sequence. end_page defaults to 10.}

\item{increase_by}{Defines by how much the number in the link should be
increased in the numerical sequence. Defaults to 1.}

\item{date_format}{A character string, defaults to "YMD". Check
\code{\link{strptime}} for valid values used to define the format of the
date that is part of the URL. Simplified formats such as the following are
also accepted: "Y" (e.g. 2022), "Ym" (2022-10), "Ymd" (e.g. 2022-10-24).
See details.}

\item{start_date}{Defaults to NULL. If given, a date, or a character vector
of length one coercible to date with \code{\link{as.Date}}. When given,
urls are built based on dates, and parameters \code{start_page}, \code{end_page}, and
\code{increase_by}, are ignored.}

\item{end_date}{Defaults to \code{Sys.Date()}. If given, a date, or a
character vector of length one coercible to date with
\code{\link{as.Date}}.}

\item{increase_date_by}{Defaults to "day". See \code{\link{seq.Date}} for
valid values.}

\item{reversed_order}{Logical, defaults to FALSE. If TRUE, the order of urls
in the output.}

\item{type}{A character vector, defaults to "index". Used for differentiating
among different types of index or links in local databases.}
}
\value{
A data frame with three columns, \code{index_id}, \code{url}, and \code{type}.
Typically, \code{url} corresponds to a vector of unique urls.
}
\description{
Convenience function typically used to generate urls to index pages listing
articles.
}
\section{Date formats}{
 It is not uncommon in particular for index pages to
include dates in the URL, along the lines of
\code{example.com/archive/2022-01-01}, \code{example.com/archive/2022-01-02}, etc. To
build such urls, \code{cas_build_urls} needs a \code{start_date} and \code{end_date}.
The formatting of the date can be defined either by providing to the
parameter \code{date_format} a string that \code{\link{strptime}} is able to
interpret directly, or a simplified string (such as "Ymd", without the
"\%"),adding a \code{date_separator} such as "-" as needed.
}

\examples{
cas_build_urls(
  url_beginning = "https://www.example.com/news/",
  start_page = 1,
  end_page = 10
)

cas_build_urls(
  url_beginning = "https://example.com/news/?skip=",
  start_page = 0,
  end_page = 100,
  increase_by = 10
)


cas_build_urls(
  url_beginning = "https://example.com/archive/",
  start_date = "2022-01-01",
  end_date = "2022-12-31",
  date_separator = "-"
) \%>\%
  head()
}
