% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_database.R
\name{cas_check_cache}
\alias{cas_check_cache}
\title{Check caching status in the current session, and override it upon request}
\usage{
cas_check_cache(cache = NULL)
}
\arguments{
\item{cache}{Defaults to NULL. If NULL, checks current cache settings. If given, returns given value, ignoring cache.}
}
\value{
Either TRUE or FALSE, depending on current cache settings.
}
\description{
Mostly used internally in functions, exported for reference.
}
\examples{
cas_check_cache()
}
\seealso{
Other database functions: 
\code{\link{cas_check_cache_folder}()},
\code{\link{cas_connect_to_cache}()},
\code{\link{cas_create_cache_folder}()},
\code{\link{cas_disable_cache}()},
\code{\link{cas_disconnect_from_cache}()},
\code{\link{cas_enable_cache}()},
\code{\link{cas_get_cache_db}()},
\code{\link{cas_set_cache_db}()},
\code{\link{cas_set_cache_folder}()}
}
\concept{database functions}
