% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_cache.R
\name{cas_check_cache_folder}
\alias{cas_check_cache_folder}
\title{Checks if cache folder exists, if not returns an informative message}
\usage{
cas_check_cache_folder()
}
\value{
If the cache folder exists, returns TRUE. Otherwise throws an error.
}
\description{
Checks if cache folder exists, if not returns an informative message
}
\examples{

# If cache folder does not exist, it throws an error
tryCatch(cas_check_cache_folder(),
  error = function(e) {
    return(e)
  }
)

# Create cache folder
cas_set_cache_folder(path = fs::path(
  tempdir(),
  "cas_cache_folder"
))
cas_create_cache_folder(ask = FALSE)

cas_check_cache_folder()
}
