% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_database.R
\name{cas_check_db_folder}
\alias{cas_check_db_folder}
\title{Checks if database folder exists, if not returns an informative message}
\usage{
cas_check_db_folder()
}
\value{
If the database folder exists, returns TRUE. Otherwise throws an error.
}
\description{
Checks if database folder exists, if not returns an informative message
}
\examples{

# If database folder does not exist, it throws an error
tryCatch(cas_check_db_folder(),
  error = function(e) {
    return(e)
  }
)

# Create database folder
cas_set_db_folder(path = fs::path(
  tempdir(),
  "cas_db_folder"
))
cas_create_db_folder(ask = FALSE)

cas_check_db_folder()
}
\seealso{
Other database functions:
\code{\link[=cas_check_use_db]{cas_check_use_db()}},
\code{\link[=cas_connect_to_db]{cas_connect_to_db()}},
\code{\link[=cas_create_db_folder]{cas_create_db_folder()}},
\code{\link[=cas_disable_db]{cas_disable_db()}},
\code{\link[=cas_disconnect_from_db]{cas_disconnect_from_db()}},
\code{\link[=cas_enable_db]{cas_enable_db()}},
\code{\link[=cas_get_db_settings]{cas_get_db_settings()}},
\code{\link[=cas_read_from_db]{cas_read_from_db()}},
\code{\link[=cas_set_db]{cas_set_db()}},
\code{\link[=cas_set_db_folder]{cas_set_db_folder()}},
\code{\link[=cas_write_to_db]{cas_write_to_db()}}
}
\concept{database functions}
