% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_check_response.R
\name{cas_check_response}
\alias{cas_check_response}
\title{Cas check httr response}
\usage{
cas_check_response(
  urls_df = NULL,
  wait = 1,
  output_only_newly_checked = TRUE,
  urls = NULL,
  index = FALSE,
  index_group = NULL,
  db_connection = NULL,
  disconnect_db = FALSE,
  check_db = TRUE,
  write_db = TRUE,
  ...
)
}
\arguments{
\item{urls_df}{A data frame with a url column, typically retrieved with \code{cas_get_urls_df()}.}

\item{wait}{Defaults to 1. Number of seconds to wait between downloading one
page and the next. Can be increased to reduce server load, or can be set to
0 when this is not an issue.}

\item{check_db}{Defaults to TRUE. If TRUE, checks if given URL has already
been checked in local database, and queries APIs only for URLs that have
not been previously checked.}

\item{write_db}{Defaults to TRUE. If TRUE, writes result to a local database.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\description{
Cas check httr response
}
