% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_database.R
\name{cas_check_use_db}
\alias{cas_check_use_db}
\title{Check caching status in the current session, and override it upon request}
\usage{
cas_check_use_db(use_db = NULL, ...)
}
\arguments{
\item{use_db}{Defaults to NULL. If NULL, checks current use_db settings. If given, returns given value, ignoring use_db.}
}
\value{
Either TRUE or FALSE, depending on current use_db settings.
}
\description{
Mostly used internally in functions, exported for reference.
}
\examples{
cas_check_use_db()
}
\seealso{
Other database functions:
\code{\link[=cas_check_db_folder]{cas_check_db_folder()}},
\code{\link[=cas_connect_to_db]{cas_connect_to_db()}},
\code{\link[=cas_create_db_folder]{cas_create_db_folder()}},
\code{\link[=cas_disable_db]{cas_disable_db()}},
\code{\link[=cas_disconnect_from_db]{cas_disconnect_from_db()}},
\code{\link[=cas_enable_db]{cas_enable_db()}},
\code{\link[=cas_get_db_settings]{cas_get_db_settings()}},
\code{\link[=cas_read_from_db]{cas_read_from_db()}},
\code{\link[=cas_set_db]{cas_set_db()}},
\code{\link[=cas_set_db_folder]{cas_set_db_folder()}},
\code{\link[=cas_write_to_db]{cas_write_to_db()}}
}
\concept{database functions}
