% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_count.R
\name{cas_count_total_words}
\alias{cas_count_total_words}
\title{Count total words in a dataset}
\usage{
cas_count_total_words(
  corpus,
  pattern = "\\\\w+",
  text = text,
  group_by = date,
  ignore_case = TRUE,
  n_column_name = n,
  locale = "en"
)
}
\arguments{
\item{corpus}{A textual corpus as a data frame.}

\item{pattern}{Defaults to pattern commonly used to count words.}

\item{text}{Defaults to \code{text}. The unquoted name of the column of the corpus
data frame to be used for matching.}

\item{group_by}{Defaults to NULL. If given, the name of one ore more columns
to be used for grouping (e.g. \code{date}, or \code{doc_id}, or \code{source}, etc.)}

\item{ignore_case}{Defaults to TRUE.}

\item{n_column_name}{Defaults to \code{n}. The unquoted name of the column to be
used for the count in the output.}

\item{locale}{Locale to be used when ignore_case is set to TRUE. Passed to
\code{stringr::str_to_lower}, defaults to "en".}
}
\description{
Count total words in a dataset
}
