% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_database.R
\name{cas_create_cache_folder}
\alias{cas_create_cache_folder}
\title{Creates the base cache folder where \code{castarter} caches data.}
\usage{
cas_create_cache_folder(ask = TRUE)
}
\arguments{
\item{ask}{Logical, defaults to TRUE. If FALSE, and cache folder does not exist, it just creates it without asking (useful for non-interactive sessions).}
}
\value{
Nothing, used for its side effects.
}
\description{
Creates the base cache folder where \code{castarter} caches data.
}
\examples{
\donttest{
if (interactive()) {
  cas_create_cache_folder()
}
}
}
\seealso{
Other database functions: 
\code{\link{cas_check_cache_folder}()},
\code{\link{cas_check_cache}()},
\code{\link{cas_connect_to_cache}()},
\code{\link{cas_disable_cache}()},
\code{\link{cas_disconnect_from_cache}()},
\code{\link{cas_enable_cache}()},
\code{\link{cas_get_cache_db}()},
\code{\link{cas_set_cache_db}()},
\code{\link{cas_set_cache_folder}()}
}
\concept{database functions}
