% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_cache.R
\name{cas_disable_cache}
\alias{cas_disable_cache}
\title{Disable caching for the current session}
\usage{
cas_disable_cache()
}
\value{
Nothing, used for its side effects.
}
\description{
Disable caching for the current session
}
\examples{
\donttest{
if (interactive()) {
  cas_disable_cache()
}
}
}
