% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_database.R
\name{cas_disconnect_from_db}
\alias{cas_disconnect_from_db}
\title{Ensure that connection to database is disconnected consistently}
\usage{
cas_disconnect_from_db(
  db_connection = NULL,
  db_type = NULL,
  use_db = NULL,
  disconnect_db = FALSE
)
}
\arguments{
\item{db_connection}{Defaults to NULL. If NULL, and database is enabled, \code{castarter} will use a local sqlite database. A custom connection to other databases can be given (see vignette \code{castarter_db_management} for details).}

\item{use_db}{Defaults to NULL. If given, it should be given either TRUE or FALSE. Typically set with \code{cas_enable_db()} or \code{cas_disable_db()}.}

\item{disconnect_db}{Defaults to TRUE. If FALSE, leaves the connection to database open.}
}
\value{
Nothing, used for its side effects.
}
\description{
Ensure that connection to database is disconnected consistently
}
\examples{
cas_disconnect_from_db()
}
\seealso{
Other database functions:
\code{\link[=cas_check_db_folder]{cas_check_db_folder()}},
\code{\link[=cas_check_use_db]{cas_check_use_db()}},
\code{\link[=cas_connect_to_db]{cas_connect_to_db()}},
\code{\link[=cas_create_db_folder]{cas_create_db_folder()}},
\code{\link[=cas_disable_db]{cas_disable_db()}},
\code{\link[=cas_enable_db]{cas_enable_db()}},
\code{\link[=cas_get_db_settings]{cas_get_db_settings()}},
\code{\link[=cas_read_from_db]{cas_read_from_db()}},
\code{\link[=cas_set_db]{cas_set_db()}},
\code{\link[=cas_set_db_folder]{cas_set_db_folder()}},
\code{\link[=cas_write_to_db]{cas_write_to_db()}}
}
\concept{database functions}
