% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_download.R
\name{cas_download}
\alias{cas_download}
\title{Downloads files systematically, and stores details about the download in a
local database}
\usage{
cas_download(
  download_df = NULL,
  index = FALSE,
  overwrite_file = FALSE,
  wait = 1,
  random = FALSE,
  file_format = "html",
  retry_times = 3,
  download_again_if_status_is_not = NULL,
  ...
)
}
\arguments{
\item{index}{Logical, defaults to FALSE. If TRUE, downloaded files will be
considered \code{index} files. If not, they will be considered \code{contents} files.
See Readme for a more extensive explanation.}

\item{overwrite_file}{Logical, defaults to FALSE. If TRUE, files are
downloaded again even if already present, overwriting previously downloaded
items.}

\item{wait}{Defaults to 1. Number of seconds to wait between downloading one
page and the next. Can be increased to reduce server load, or can be set to
0 when this is not an issue.}

\item{random}{Defaults to FALSE. If TRUE, the download order is randomised.
If a numeric is given, the download order is randomised and at most the
given number of items is downloaded.}

\item{retry_times}{Defaults to 3. Number of times to retry download in case
of errors.}

\item{...}{Passed to \code{cas_get_db_file()}.}

\item{urls_df}{A data frame with at least two columns named \code{id} and \code{url}.
Typically generated with \code{cas_build_urls()} for index files. If a character
vector is given instead, identifiers will be given automatically.}
}
\description{
Downloads files systematically, and stores details about the download in a
local database
}
