% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_download_httr.R
\name{cas_download_httr}
\alias{cas_download_httr}
\title{Downloads one file at a time with httr}
\usage{
cas_download_httr(
  download_df = NULL,
  index = FALSE,
  index_group = NULL,
  overwrite_file = FALSE,
  ignore_id = TRUE,
  wait = 1,
  create_folder_if_missing = NULL,
  pause_base = 2,
  pause_cap = 256,
  pause_min = 4,
  terminate_on = NULL,
  retry_times = 3,
  db_connection = NULL,
  disconnect_db = FALSE,
  sample = FALSE,
  file_format = "html",
  user_agent = NULL,
  download_again = FALSE,
  download_again_if_status_is_not = NULL,
  ...
)
}
\arguments{
\item{download_df}{A data frame with four columns: \code{id}, \code{url}, \code{path}, \code{type}.}

\item{index}{Logical, defaults to FALSE. If TRUE, downloaded files will be
considered \code{index} files. If not, they will be considered \code{contents} files.
See Readme for a more extensive explanation.}

\item{overwrite_file}{Logical, defaults to \code{FALSE}.}

\item{wait}{Defaults to 1. Number of seconds to wait between downloading one
page and the next. Can be increased to reduce server load, or can be set to
0 when this is not an issue.}

\item{retry_times}{Defaults to 3. Number of times to retry download in case
of errors.}

\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{disconnect_db}{Defaults to \code{FALSE}. If \code{FALSE}, leaves the connection
to database open.}

\item{sample}{Defaults to FALSE. If TRUE, the download order is randomised.
If a numeric is given, the download order is randomised and at most the
given number of items is downloaded.}

\item{file_format}{Defaults to \code{html}. Used for storing files in dedicated
folders, but also for determining processing options. For example, if a
sitemap is downloaded as an index with \code{file_format} set to xml, it will be
processed accordingly. If it is stored as xml.gz, it will be automatically
decompressed for correct processing.}

\item{user_agent}{Defaults to NULL. If given, passed to download method.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\value{
Invisibly returns the full \code{httr} response.
}
\description{
Mostly used internally by \code{\link[=cas_download]{cas_download()}}.
}
