% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_database.R
\name{cas_enable_db}
\alias{cas_enable_db}
\title{Enable caching for the current session}
\usage{
cas_enable_db(SQLite = TRUE)
}
\arguments{
\item{SQLite}{Logical, defaults to TRUE. Set to FALSE to use custom database options. See \code{cas_set_db()} for details.}
}
\value{
Nothing, used for its side effects.
}
\description{
Enable caching for the current session
}
\examples{
\donttest{
if (interactive()) {
  cas_enable_db()
}
}
}
\seealso{
Other database functions: 
\code{\link{cas_check_db_folder}()},
\code{\link{cas_check_use_db}()},
\code{\link{cas_connect_to_db}()},
\code{\link{cas_create_db_folder}()},
\code{\link{cas_disable_db}()},
\code{\link{cas_disconnect_from_db}()},
\code{\link{cas_get_db_settings}()},
\code{\link{cas_set_db_folder}()},
\code{\link{cas_set_db}()}
}
\concept{database functions}
