% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_extract.R
\name{cas_extract}
\alias{cas_extract}
\title{Extract fields and contents from downloaded files}
\usage{
cas_extract(
  extractors,
  id = NULL,
  index = FALSE,
  db_connection = NULL,
  file_format = "html",
  random = FALSE,
  write_to_db = TRUE,
  keep_if_status = 200,
  ...
)
}
\arguments{
\item{extractors}{A named list of functions. See examples for details.}

\item{index}{Logical, defaults to FALSE. If TRUE, downloaded files will be
considered \code{index} files. If not, they will be considered \code{contents} files.
See Readme for a more extensive explanation.}

\item{random}{Defaults to FALSE. If TRUE, the download order is randomised.
If a numeric is given, the download order is randomised and at most the
given number of items is downloaded.}

\item{keep_if_status}{Defaults to 200. Keep only if recorded download status matches the given status.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\description{
Extract fields and contents from downloaded files
}
