% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_extract.R
\name{cas_extract}
\alias{cas_extract}
\title{Extract fields and contents from downloaded files}
\usage{
cas_extract(
  extractors,
  index = FALSE,
  db_connection = NULL,
  file_format = "html",
  write_to_db = TRUE,
  ...
)
}
\arguments{
\item{extractors}{A named list of functions. See examples for details.}

\item{index}{Logical, defaults to FALSE. If TRUE, downloaded files will be
considered \code{index} files. If not, they will be considered \code{contents} files.
See Readme for a more extensive explanation.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\description{
Extract fields and contents from downloaded files
}
