% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_extract_links.R
\name{cas_extract_links}
\alias{cas_extract_links}
\title{Extract direct links to individual content pages from index pages}
\usage{
cas_extract_links(
  id = NULL,
  batch = "latest",
  domain = NULL,
  index = TRUE,
  index_group = NULL,
  output_index = FALSE,
  output_index_group = NULL,
  include_when = NULL,
  exclude_when = NULL,
  container = NULL,
  container_class = NULL,
  container_id = NULL,
  custom_xpath = NULL,
  custom_css = NULL,
  match = NULL,
  min_length = NULL,
  max_length = NULL,
  attribute_type = "href",
  append_string = NULL,
  remove_string = NULL,
  write_to_db = FALSE,
  file_format = "html",
  keep_only_within_domain = TRUE,
  sample = FALSE,
  check_previous = TRUE,
  check_again = FALSE,
  encoding = "UTF-8",
  reverse_order = FALSE,
  db_connection = NULL,
  disconnect_db = TRUE,
  ...
)
}
\arguments{
\item{id}{Defaults to NULL. If provided, it should be a vector of integers.
Only html files corresponding to given id will be processed.}

\item{domain}{Defaults to NULL. Web domain of the website. It is added at the
beginning of each link found. If links in the page already include the full
web address this should be ignored.}

\item{index}{Logical, defaults to FALSE. If TRUE, downloaded files will be
considered \code{index} files. If not, they will be considered \code{contents} files.
See Readme for a more extensive explanation.}

\item{index_group}{Defaults to NULL. If TRUE, only items in the given index
group will be parsed.}

\item{output_index}{Defaults to FALSE. If FALSE, new links are added to the
contents table. If TRUE, the links extracted will be stored again as
index, using \code{output_index_group} as \code{index_group}.}

\item{output_index_group}{Defaults to NULL. Relevant only when \code{output_index}
is set to TRUE. Used to store new index urls in the database with reference
to the appropriate group.}

\item{include_when}{Part of URL found only in links of individual articles to
be downloaded. If more than one provided, it includes all links that
contains either of the strings provided.}

\item{exclude_when}{If an URL includes this string, it is excluded from the
output. One or more strings may be provided.}

\item{container}{Defaults to \code{NULL}. Type of html container from where links
are to be extracted, such as "div", "ul", and others. Either
\code{container_class} or \code{container_id} must also be provided.}

\item{container_class}{Defaults to \code{NULL}. If provided, also \code{container} must
be given (and \code{container_id} must be \code{NULL}). Only text found inside the
provided combination of container/class will be extracted.}

\item{container_id}{Defaults to \code{NULL}. If provided, also \code{container} must be
given (and \code{container_id} must be \code{NULL}). Only text found inside the
provided combination of container/class will be extracted.}

\item{custom_xpath}{Defaults to \code{NULL}. If given, all other parameters are
ignored and given Xpath used instead.}

\item{match}{Defaults to NULL. Used when extracting json files. Name of
property from where url is to be extracted. N.B. Only partly implemented,
please report issues along with specific example where it emerged.}

\item{min_length}{If a link is shorter than the number of characters given in
min_length, it is excluded from the output.}

\item{max_length}{If a link is longer than the number of characters given in
max_length, it is excluded from the output.}

\item{attribute_type}{Defaults to "href". Type of attribute to extract from
links.}

\item{append_string}{If provided, appends given string to the extracted
articles. Typically used to create links for print or mobile versions of
the extracted page.}

\item{remove_string}{If provided, remove given string (or strings) from
links.}

\item{write_to_db}{Logical, defaults to FALSE. If TRUE stored newly extracted
links in the database, associates each of them with an id, and records the
source for each link.}

\item{keep_only_within_domain}{Logical, defaults to TRUE. If TRUE, and domain
given, links to external websites are dropped.}

\item{check_previous}{Defaults to TRUE. If TRUE, checks if newly found links
are previously stored in database, and if they are, it discards them. If
FALSE, and \code{write_to_db} is also set to FALSE, it does not check for
previously stored links.}

\item{check_again}{Defaults to FALSE. If FALSE, files from where are at least
a link has been extracted are not re-processed. If TRUE, they are processed
again. By default, only new links are then actually included in the output
or stored in the local database.}

\item{reverse_order}{Logical, defaults to FALSE. If TRUE, index files are
processed in reverse order of \code{id} and \code{batch}, which may give more
meaningful order to content id. The difference is ultimately cosmetic, and
has no substantive impact either way.}

\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{disconnect_db}{Defaults to \code{FALSE}. If \code{FALSE}, leaves the connection
to database open.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\value{
A data frame.
}
\description{
Extract direct links to individual content pages from index pages
}
\examples{
\dontrun{
links <- cas_extract_links(domain = "http://www.example.com/")
}
}
