% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_extract_links.R
\name{cas_extract_links}
\alias{cas_extract_links}
\title{Extract direct links to individual content pages from index pages}
\usage{
cas_extract_links(
  id = NULL,
  domain = NULL,
  index = TRUE,
  include_when = NULL,
  exclude_when = NULL,
  container = NULL,
  container_class = NULL,
  container_id = NULL,
  min_length = NULL,
  max_length = NULL,
  attribute_type = "href",
  append_string = NULL,
  remove_string = NULL,
  write_to_db = TRUE,
  file_format = "html",
  random = FALSE,
  encoding = "UTF-8",
  ...
)
}
\arguments{
\item{id}{Defaults to NULL. If provided, it should be a vector of integers. Only html files corresponding to given id in the relevant will be processed.}

\item{domain}{Defaults to "". Web domain of the website. It is added at the beginning of each link found. If links in the page already include the full web address this should be ignored.}

\item{include_when}{Part of URL found only in links of individual articles to be downloaded. If more than one provided, it includes all links that contains either of the strings provided.}

\item{exclude_when}{If an URL includes this string, it is excluded from the output. One or more strings may be provided.}

\item{container}{Defaults to NULL. Type of html container from where links are to be extracted, such as "div", "ul", and others. Either \code{container_class} or \code{container_id} must also be provided.}

\item{container_class}{Defaults to NULL. If provided, also \code{container} must be given (and \code{container_id} must be NULL). Only text found inside the provided combination of container/class will be extracted.}

\item{container_id}{Defaults to NULL. If provided, also \code{container} must be given (and \code{container_id} must be NULL). Only text found inside the provided combination of container/class will be extracted.}

\item{min_length}{If a link is shorter than the number of characters given in min_length, it is excluded from the output.}

\item{max_length}{If a link is longer than the number of characters given in max_length, it is excluded from the output.}

\item{attribute_type}{Defaults to "href". Type of attribute to extract from links.}

\item{append_string}{If provided, appends given string to the extracted articles. Typically used to create links for print or mobile versions of the extracted page.}

\item{remove_string}{If provided, remove given string (or strings) from links.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\value{
A data frame.
}
\description{
Extract direct links to individual content pages from index pages
}
\examples{
\dontrun{
links <- cas_extract_links(domain = "http://www.example.com/")
}
}
