% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_extract.R
\name{cas_extract_script}
\alias{cas_extract_script}
\title{Extracts scripts from an html page}
\usage{
cas_extract_script(
  html_document,
  script_type = NULL,
  match = NULL,
  accessors = NULL
)
}
\arguments{
\item{html_document}{An html document parsed with \code{xml2::read_html()}.}

\item{script_type}{Defaults to NULL. Type of script. Common script types
include \code{application/ld+json}, \code{text/template}, etc.}

\item{match}{Default to NULL. If given, used to filter extracted scripts.
Must be a named vector in the format \verb{c(}@type\verb{ = "NewsArticle")} for a
script of type "NewsArticle".}

\item{accessors}{Defaults to NULL. If given, a vector of accessors passed to
\code{purrr::pluck} in order to extract sub-components of the list resulting
from reading the with \code{jsonlite} the result of the previous steps and
filter.}
}
\value{
May return a list or a character vector. If no match is found, returns \code{NA_character_}
}
\description{
Extracts scripts from an html page
}
\examples{
\dontrun{
if (interactive()) {
  url <- "https://www.digi24.ro/stiri/externe/casa-alba-pune-capat-isteriei-globale-nu-exista-indicii-ca-obiectele-zburatoare-doborate-de-rachetele-sua-ar-fi-extraterestre-2250863"

  html_document <- rvest::read_html(x = url)

  cas_extract_script(
    html_document = html_document,
    script_type = "application/ld+json"
  )

  # get date published
  cas_extract_script(
    html_document = html_document,
    script_type = "application/ld+json",
    match = c(`@type` = "NewsArticle"),
    accessors = "datePublished"
  )

  # get title
  cas_extract_script(
    html_document = html_document,
    script_type = "application/ld+json",
    match = c(`@type` = "NewsArticle"),
    accessors = "headline"
  )

  # get nested element, e.g. url of the logo of the publisher

  cas_extract_script(
    html_document = html_document,
    script_type = "application/ld+json",
    match = c(`@type` = "NewsArticle"),
    accessors = cas("publisher", "logo", "url")
  )
}
}
}
