% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_find_extractor.R
\name{cas_find_extractor}
\alias{cas_find_extractor}
\title{Facilitate finding extractors, typically  to be used with
\code{cas_extract_html()}}
\usage{
cas_find_extractor(
  html_document,
  pattern,
  containers = c("h1", "h2", "h3", "h4", "span", "td", "div")
)
}
\arguments{
\item{html_document}{An html document parsed with \code{xml2::read_html()} or
\code{rvest::read_html()}.}

\item{pattern}{A text string to be matched.}

\item{containers}{Containers to be parsed for best matches. By default:
\verb{c("h1", "h2", "h3", "h4", "span", "td", div", "span")}. The order matters,
as results are returned in this order (e.g. if a match of the same length
is found both in a "h1" and in a "div", "h1" is returned first).}
}
\value{
A data frame list with container and class or id of values that
should work if passed to \code{cas_extract_html()}.
}
\description{
This may or may not work, but it may be worth giving this a quick a try
before looking for alternatives. The parameters returned first should work
best.
}
\examples{
\dontrun{
if (interactive) {
  # not ideal example, but you'll get the gist
  library("castarter")
  url <- "https://www.nasa.gov/news-release/nasa-sets-coverage-for-roscosmos-spacewalk-outside-space-station/"

  html_page <- rvest::read_html(url)

  cas_find_extractor(
    html_document = html_page,
    pattern = "NASA Sets Coverage for Roscosmos Spacewalk Outside Space Station"
  )

  cas_find_extractor(
    html_document = html_page,
    pattern = "Oct 23, 2023"
  )

  cas_find_extractor(
    html_document = html_page,
    pattern = "Roxana Bardan"
  )

  cas_find_extractor(
    html_document = html_page,
    pattern = "RELEASE"
  )
}
}

}
