% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_cache.R
\name{cas_get_cache_db}
\alias{cas_get_cache_db}
\title{Get database connection settings from the environment}
\usage{
cas_get_cache_db()
}
\value{
A list with all database parameters as stored in environment variables.
}
\description{
Typically set with \code{cas_set_cache_db()}
}
\examples{

cas_get_cache_db()
}
\seealso{
Other caching functions: 
\code{\link{cas_check_cache_folder}()},
\code{\link{cas_check_cache}()},
\code{\link{cas_connect_to_cache}()},
\code{\link{cas_create_cache_folder}()},
\code{\link{cas_disable_cache}()},
\code{\link{cas_disconnect_from_cache}()},
\code{\link{cas_enable_cache}()},
\code{\link{cas_set_cache_db}()},
\code{\link{cas_set_cache_folder}()}
}
\concept{caching functions}
