% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_database.R, R/cas_files.R
\name{cas_get_db}
\alias{cas_get_db}
\title{Get database connection settings from the environment}
\usage{
cas_get_db(base_folder = NULL, project = NULL, website = NULL)

cas_get_db(base_folder = NULL, project = NULL, website = NULL)
}
\arguments{
\item{base_folder}{Defaults to NULL, can be set once per session with
\code{cas_set_options()}. A path to a location used for storing html and other
project files. If the folder does not exist, it will be created. If not
given, and not previously set as environment variable, defaults to
\code{castarter_data}.}

\item{project}{Defaults to NULL. Project name, can be set once per session
with \code{cas_set_options()}. This will be used as first level folder and may
be used elsewhere to describe the dataset.}

\item{website}{Defaults to NULL. Website name, can be set once per session
with \code{cas_set_options()}. This will be used as a second level folder and
may be used elsewhere to describe the dataset.}
}
\value{
A list with all database parameters as stored in environment variables.


}
\description{
Typically set with \code{cas_set_db()}
}
\examples{

cas_get_db()

cas_get_db(
  base_folder = fs::path_temp(),
  project = "example_project",
  website = "example_website"
)
}
\seealso{
Other database functions: 
\code{\link{cas_check_db_folder}()},
\code{\link{cas_check_use_db}()},
\code{\link{cas_connect_to_db}()},
\code{\link{cas_create_db_folder}()},
\code{\link{cas_disable_db}()},
\code{\link{cas_disconnect_from_db}()},
\code{\link{cas_enable_db}()},
\code{\link{cas_set_db_folder}()},
\code{\link{cas_set_db}()}
}
\concept{database functions}
