% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_database.R
\name{cas_get_db_settings}
\alias{cas_get_db_settings}
\title{Get database connection settings from the environment}
\usage{
cas_get_db_settings()
}
\value{
A list with all database parameters as stored in environment variables.
}
\description{
Typically set with \code{cas_set_db()}
}
\examples{

cas_get_db_settings()
}
\seealso{
Other database functions: 
\code{\link{cas_check_db_folder}()},
\code{\link{cas_check_use_db}()},
\code{\link{cas_connect_to_db}()},
\code{\link{cas_create_db_folder}()},
\code{\link{cas_disable_db}()},
\code{\link{cas_disconnect_from_db}()},
\code{\link{cas_enable_db}()},
\code{\link{cas_read_from_db}()},
\code{\link{cas_set_db_folder}()},
\code{\link{cas_set_db}()},
\code{\link{cas_write_to_db}()}
}
\concept{database functions}
