% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_options.R
\name{cas_get_options}
\alias{cas_get_options}
\title{Retrieve options}
\usage{
cas_get_options(base_folder = NULL, project = NULL, website = NULL)
}
\arguments{
\item{base_folder}{A path to a location used for storing html and other project files. If the folder does not exist, it will be created. Defaults to \code{castarter_data}}

\item{project}{Project name. This will be used as first level folder and may be used elsewhere to describe the dataset.}

\item{website}{Website name. This will be used as a second level folder and may be used elsewhere to describe the dataset.}
}
\value{
A list object with the previously set options,
}
\description{
Retrieve options
}
\examples{

cas_get_options()

}
