% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_options.R
\name{cas_get_options}
\alias{cas_get_options}
\title{Get key project parameters that determine the folder used for storing project files}
\usage{
cas_get_options(
  project = NULL,
  website = NULL,
  use_db = NULL,
  base_folder = NULL,
  db_type = NULL,
  db_folder = NULL
)
}
\arguments{
\item{project}{Defaults to NULL. Project name, can be set once per session
with \code{\link[=cas_set_options]{cas_set_options()}}. This will be used as first level folder and may
be used elsewhere to describe the dataset.}

\item{website}{Defaults to NULL. Website name, can be set once per session
with \code{\link[=cas_set_options]{cas_set_options()}}. This will be used as a second level folder and
may be used elsewhere to describe the dataset.}

\item{use_db}{Defaults to TRUE. If TRUE, stores information about the
download process and extracted text in a local database.}

\item{base_folder}{Defaults to NULL, can be set once per session with
\code{\link[=cas_set_options]{cas_set_options()}}. A path to a location used for storing html and other
project files. If the folder does not exist, it will be created. If not
given, and not previously set as environment variable, defaults to
\code{castarter_data}.}

\item{db_folder}{Defaults to NULL. can be set once per session with
\code{\link[=cas_set_options]{cas_set_options()}} or \code{\link[=cas_set_db_folder]{cas_set_db_folder()}}. A path to a location used for
storing the database. If not given, and not previously set as environment
variable, defaults to \code{castarter_data}.}
}
\value{
A list object with the given or previously set options.
}
\description{
Get key project parameters that determine the folder used for storing project files
}
\seealso{
Other settings: 
\code{\link{cas_set_options}()}
}
\concept{settings}
