% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_get_path_to_files.R
\name{cas_get_path_to_files}
\alias{cas_get_path_to_files}
\title{Get path to locally downloaded files}
\usage{
cas_get_path_to_files(
  urls = NULL,
  index = FALSE,
  custom_folder = NULL,
  custom_path = NULL,
  file_format = "html",
  db_connection = NULL,
  ...
)
}
\description{
This function relies on data stored in the database.
}
