% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_get_path_to_files.R
\name{cas_get_path_to_files}
\alias{cas_get_path_to_files}
\title{Get path to locally downloaded files}
\usage{
cas_get_path_to_files(
  urls = NULL,
  id = NULL,
  index = FALSE,
  index_group = NULL,
  custom_folder = NULL,
  custom_path = NULL,
  file_format = "html",
  sample = FALSE,
  db_connection = NULL,
  db_folder = NULL,
  ...
)
}
\arguments{
\item{index}{Logical, defaults to FALSE. If TRUE, downloaded files will be
considered \code{index} files. If not, they will be considered \code{contents} files.
See Readme for a more extensive explanation.}

\item{sample}{Defaults to FALSE. If TRUE, the download order is randomised.
If a numeric is given, the download order is randomised and at most the
given number of items is downloaded.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\description{
This function relies on data stored in the database.
}
