% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_get_sitemap.R
\name{cas_get_sitemap}
\alias{cas_get_sitemap}
\title{Check for availability of sitemap.xml}
\usage{
cas_get_sitemap(domain, sitemap_url = NULL)
}
\arguments{
\item{sitemap_url}{}
}
\value{
A data frame, including a \code{sitemap_url} column, the response as an
httr2 object, and body of the xml.
}
\description{
Searches in default location and then in robots.txt and return url
}
\examples{
if (interactive) {
cas_get_sitemap("https://www.europeandatajournalism.eu/")
}
}
