% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_ia_check.R
\name{cas_ia_check}
\alias{cas_ia_check}
\title{Gets an Archive.org Wayback Machine URL}
\usage{
cas_ia_check(url, wait = 1)
}
\arguments{
\item{url}{A charachter vector of length one, a url.}

\item{wait}{Defaults to 1. Number of seconds to wait between downloading one
page and the next. Can be increased to reduce server load, or can be set to
0 when this is not an issue.}
}
\value{
A url linking to the version on the Internet Archive
}
\description{
For details on API access to the Wayback Machine see:
https://archive.org/help/wayback_api.php
}
\details{
For an R package facilitating more extensive interaction with the API, see:
https://github.com/hrbrmstr/wayback
}
\examples{
\dontrun{

}
}
