% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_read_corpus.R
\name{cas_read_corpus}
\alias{cas_read_corpus}
\title{Read datasets created with \code{cas_write_dataset}}
\usage{
cas_read_corpus(file_format = "parquet", ...)
}
\arguments{
\item{file_format}{Defaults to "parquet". Currently, other options are not implemented.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\value{
A dataset as \code{ArrowObject}
}
\description{
Read datasets created with \code{cas_write_dataset}
}
\examples{
\dontrun{
cas_read_corpus()
}
}
