% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_check_response.R
\name{cas_read_db_response}
\alias{cas_read_db_response}
\title{Check response type of URLs as stored in the local database.}
\usage{
cas_read_db_response(db_connection = NULL, db_folder = NULL, ...)
}
\arguments{
\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\description{
Data stored locally typically by relying on \code{cas_check_response()}.
}
