% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_database.R
\name{cas_read_from_db}
\alias{cas_read_from_db}
\title{Reads data from local database}
\usage{
cas_read_from_db(
  table,
  use_db = NULL,
  db_connection = NULL,
  disconnect_db = TRUE
)
}
\arguments{
\item{table}{Name of the table. See readme for details.}

\item{use_db}{Defaults to NULL. If given, it should be given either TRUE or FALSE. Typically set with \code{cas_enable_db()} or \code{cas_disable_db()}.}

\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database. If given, must be a connection object or a list with relevant connection settings (see example).}

\item{disconnect_db}{Defaults to TRUE. If FALSE, leaves the connection to database open.}
}
\description{
Reads data from local database
}
\examples{
cas_set_options(base_folder = fs::path(tempdir(), "R", "castarter_data"),
                project = "example_project",
                website = "example_website"
)
cas_enable_db()


urls_df <- cas_build_urls(
  url_beginning = "https://www.example.com/news/",
  start_page = 1,
  end_page = 10
)

cas_write_to_db(df = urls_df,
                table = "index_id")

cas_read_from_db(table = "index_id")
}
\seealso{
Other database functions: 
\code{\link{cas_check_db_folder}()},
\code{\link{cas_check_use_db}()},
\code{\link{cas_connect_to_db}()},
\code{\link{cas_create_db_folder}()},
\code{\link{cas_disable_db}()},
\code{\link{cas_disconnect_from_db}()},
\code{\link{cas_enable_db}()},
\code{\link{cas_get_db_settings}()},
\code{\link{cas_set_db_folder}()},
\code{\link{cas_set_db}()},
\code{\link{cas_write_to_db}()}
}
\concept{database functions}
