% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_database.R
\name{cas_set_db}
\alias{cas_set_db}
\title{Set database connection settings for the session}
\usage{
cas_set_db(
  db_settings = NULL,
  driver = NULL,
  host = NULL,
  port,
  database,
  user,
  pwd
)
}
\arguments{
\item{db_settings}{A list of database connection settings (see example)}

\item{driver}{A database driver. Common database drivers include \code{MySQL}, \code{PostgreSQL}, and \code{MariaDB}. See \code{unique(odbc::odbcListDrivers()[[1]])} for a list of locally available drivers.}

\item{host}{Host address, e.g. "localhost".}

\item{port}{Port to use to connect to the database.}

\item{database}{Database name.}

\item{user}{Database user name.}

\item{pwd}{Password for the database user.}
}
\value{
A list with all given parameters (invisibly).
}
\description{
Set database connection settings for the session
}
\examples{
\donttest{
if (interactive()) {
  # Settings can be provided either as a list
  db_settings <- list(
    driver = "MySQL",
    host = "localhost",
    port = 3306,
    database = "castarter",
    user = "secret_username",
    pwd = "secret_password"
  )

  cas_set_db(db_settings)

  # or as parameters

  cas_set_db(
    driver = "MySQL",
    host = "localhost",
    port = 3306,
    database = "castarter",
    user = "secret_username",
    pwd = "secret_password"
  )
}
}
}
\seealso{
Other database functions:
\code{\link[=cas_check_db_folder]{cas_check_db_folder()}},
\code{\link[=cas_check_use_db]{cas_check_use_db()}},
\code{\link[=cas_connect_to_db]{cas_connect_to_db()}},
\code{\link[=cas_create_db_folder]{cas_create_db_folder()}},
\code{\link[=cas_disable_db]{cas_disable_db()}},
\code{\link[=cas_disconnect_from_db]{cas_disconnect_from_db()}},
\code{\link[=cas_enable_db]{cas_enable_db()}},
\code{\link[=cas_get_db_settings]{cas_get_db_settings()}},
\code{\link[=cas_read_from_db]{cas_read_from_db()}},
\code{\link[=cas_set_db_folder]{cas_set_db_folder()}},
\code{\link[=cas_write_to_db]{cas_write_to_db()}}
}
\concept{database functions}
