% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_options.R
\name{cas_set_options}
\alias{cas_set_options}
\title{Set folder for caching data}
\usage{
cas_set_options(base_folder = NULL, project = NULL, website = NULL)
}
\arguments{
\item{path}{A path to a location used for caching data. If the folder does not exist, it will be created.}
}
\value{
The path to the caching folder, if previously set; the same path as given to the function; or the default, \code{sn_data} is none is given.
}
\description{
Consider using a folder out of your current project directory, e.g. \code{sn_set_cache_folder("~/R/sn_data/")}: you will be able to use the same cache in different projects, and prevent cached files from being sync-ed if you use services such as Nextcloud or Dropbox.
}
\examples{
\dontrun{
sn_set_cache_folder("~/R/sn_data/")
}
}
