% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_show_barchart.R
\name{cas_show_gg_base}
\alias{cas_show_gg_base}
\title{Creates base ggplot2 object to be used by ggplot or ggiraph}
\usage{
cas_show_gg_base(
  count_df,
  group_by = date,
  n_column_name = n,
  string_column_name = string,
  group_as_factor = FALSE
)
}
\arguments{
\item{group_by}{Defaults to NULL. If given, the unquoted name of the column to be used for grouping (e.g. date, or doc_id, or source, etc.)}

\item{n_column_name}{Defaults to 'n'. The unquoted name of the column to be used for the count in the output.}

\item{string_column_name}{Defaults to 'string'. The unquoted name of the column to be used for the word in the output.}

\item{group_as_factor}{Defaults to FALSE. If TRUE, the grouping column is forced into a factor, otherwise it is kept in its current format (e.g. date, or numeric).}
}
\value{
A ggplot2 object with aesthetics set, but no geometry.
}
\description{
Creates base ggplot2 object to be used by ggplot or ggiraph
}
