% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_show_ts.R
\name{cas_show_ts_dygraph}
\alias{cas_show_ts_dygraph}
\title{Create dygraphs based on a data frame typically generated with cas_count()}
\usage{
cas_show_ts_dygraph(
  count_df,
  date_column_name = date,
  n_column_name = n,
  string_column_name = string,
  range_selector = TRUE
)
}
\arguments{
\item{count_df}{}
}
\description{
Create dygraphs based on a data frame typically generated with cas_count()
}
\examples{

count_df <- castarter2::cas_count(
  corpus = castarter2::cas_demo_corpus,
  words = c("russia", "moscow")
) \%>\%
  cas_summarise(before = 15, after = 15)
cas_show_ts_dygraph(count_df)
}
