% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_wp_api.R
\name{cas_wp_api}
\alias{cas_wp_api}
\title{Retrieve data through Wordpress API}
\usage{
cas_wp_api(
  api_base_url,
  api_endpoint = "posts",
  per_page = 10,
  order_by = c("id", "date", "relevance", "include", "title", "slug"),
  order = c("desc", "asc"),
  output_only_newly_checked = FALSE,
  output_only_cached = FALSE,
  db_connection = NULL,
  disconnect_db = FALSE,
  check_db = TRUE,
  write_db = TRUE,
  ignore_ssl_certificates = FALSE,
  user_agent = NULL,
  wait = 1,
  ...
)
}
\arguments{
\item{api_base_url}{Most correspond to the base URL of a Wordpress API
endpoint.}

\item{api_endpoint}{The endpoint, defaults to "posts". Endpoints can be
specific to each Wordpress deployment or be offered by plugins, but common
endpoints include, for example, "author", "categories", "pages", etc.}

\item{per_page}{Defaults to 10. By default, this is capped at 100 by
Wordpress server-side.}

\item{order_by}{Defaults to "id", available values depend on the given
endpoint. Common options include "date".}

\item{order}{Defaults to "desc", for descending order. The only other valid
value is "asc", for ascending order.}

\item{output_only_cached}{Defaults to \code{FALSE}. If \code{TRUE}, only previously
cached responses are kept.}

\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{check_db}{Defaults to \code{TRUE}. If \code{TRUE}, checks if given URL has already
been checked in local database, and queries APIs only for URLs that have
not been previously checked.}

\item{write_db}{Defaults to \code{TRUE}. If \code{TRUE}, writes result to a local database.}

\item{wait}{Defaults to 1. Number of seconds to wait between downloading one
page and the next. Can be increased to reduce server load, or can be set to
0 when this is not an issue.}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\value{
A data frame, likely including list columns, including API response.
}
\description{
See also
\href{https://developer.wordpress.org/rest-api/using-the-rest-api/}{Wordpress
REST API official documentation.}
}
