% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_database.R
\name{cas_write_to_db}
\alias{cas_write_to_db}
\title{Generic function for writing to database}
\usage{
cas_write_to_db(
  df,
  table,
  use_db = NULL,
  overwrite = FALSE,
  db_connection = NULL,
  disconnect_db = TRUE
)
}
\arguments{
\item{df}{A data frame. Must correspond with the type of data expected for each table.}

\item{table}{Name of the table. See readme for details.}

\item{use_db}{Defaults to NULL. If given, it should be given either TRUE or FALSE. Typically set with \code{cas_enable_db()} or \code{cas_disable_db()}.}

\item{overwrite}{Logical, defaults to FALSE. If TRUE, checks if matching data are previously held in the table and overwrites them. This should be used with caution, as it may overwrite completely the selected table.}

\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database. If given, must be a connection object or a list with relevant connection settings (see example).}

\item{disconnect_db}{Defaults to TRUE. If FALSE, leaves the connection to database open.}
}
\value{
If successful, returns silently the same data frame provided as input and written to the database. Returns silently NULL, if nothing is added, e.g. because \code{use_db} is set to FALSE.
}
\description{
Generic function for writing to database
}
\seealso{
Other database functions: 
\code{\link{cas_check_db_folder}()},
\code{\link{cas_check_use_db}()},
\code{\link{cas_connect_to_db}()},
\code{\link{cas_create_db_folder}()},
\code{\link{cas_disable_db}()},
\code{\link{cas_disconnect_from_db}()},
\code{\link{cas_enable_db}()},
\code{\link{cas_get_db_settings}()},
\code{\link{cas_read_from_db}()},
\code{\link{cas_set_db_folder}()},
\code{\link{cas_set_db}()}
}
\concept{database functions}
