% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_database.R
\name{cas_write_to_db}
\alias{cas_write_to_db}
\title{Generic function for writing to database}
\usage{
cas_write_to_db(df, table, overwrite = FALSE, db_connection = NULL, ...)
}
\arguments{
\item{df}{A data frame. Must correspond with the type of data expected for each table.}

\item{table}{Name of the table. See readme for details.}

\item{overwrite}{Logical, defaults to FALSE. If TRUE, checks if matching data are previously held in the table and overwrites them. This should be used with caution, as it may overwrite completely the selected table.}

\item{db_connection}{Defaults to NULL. If NULL, uses local SQLite database.
If given, must be a connection object or a list with relevant connection
settings (see example).}

\item{...}{Passed to \code{cas_get_db_file()}.}
}
\value{
If successful, returns silently the same data frame provided as input and written to the database. Returns silently NULL, if nothing is added, e.g. because \code{use_db} is set to FALSE.
}
\description{
Generic function for writing to database
}
\examples{

cas_set_options(
  base_folder = fs::path(tempdir(), "R", "castarter_data"),
  project = "example_project",
  website = "example_website"
)
cas_enable_db()


urls_df <- cas_build_urls(
  url = "https://www.example.com/news/",
  start_page = 1,
  end_page = 10
)

cas_write_to_db(
  df = urls_df,
  table = "index_id"
)
}
\seealso{
Other database functions: 
\code{\link{cas_check_db_folder}()},
\code{\link{cas_check_use_db}()},
\code{\link{cas_connect_to_db}()},
\code{\link{cas_create_db_folder}()},
\code{\link{cas_disable_db}()},
\code{\link{cas_disconnect_from_db}()},
\code{\link{cas_enable_db}()},
\code{\link{cas_get_db_settings}()},
\code{\link{cas_read_from_db}()},
\code{\link{cas_set_db_folder}()},
\code{\link{cas_set_db}()}
}
\concept{database functions}
