% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_download.R
\name{cass_download_httr}
\alias{cass_download_httr}
\title{Downloads files, and stores details about the download in a local database}
\usage{
cass_download_httr(
  download_df,
  type = "contents",
  overwrite_file = FALSE,
  use_db = NULL,
  db_connection = NULL,
  disconnect_db = TRUE
)
}
\arguments{
\item{download_df}{A data frame with one row, and four columns: \code{id}, \code{url}, \code{path}, \code{type}.}

\item{type}{Accepted values are either "contents" (default) oe "index".}

\item{overwrite}{Logical, defaults to FALSE.}
}
\value{
Invisibly returns the full \code{httr} response.
}
\description{
Downloads files, and stores details about the download in a local database
}
