% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_download.R
\name{cass_get_files_to_download}
\alias{cass_get_files_to_download}
\title{Create a data frame with not yet downloaded files}
\usage{
cass_get_files_to_download(
  url_df = NULL,
  type = "contents",
  custom_folder = NULL,
  custom_path = NULL,
  file_format = "html",
  project = NULL,
  website = NULL,
  base_folder = NULL,
  use_db = NULL,
  db_connection = NULL,
  disconnect_db = TRUE
)
}
\arguments{
\item{url_df}{Defaults to NULL. If given, with at least two columns named \code{id} and \code{url}. If not given, an attempt will be made to load it from the local database.}
}
\value{
A data frame with four columns: \code{id}, \code{url}, \code{path} and \code{type}
}
\description{
Create a data frame with not yet downloaded files
}
