% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cass_highlight.R
\name{cass_highlight}
\alias{cass_highlight}
\title{Takes a character vector and returns it with matches of pattern wrapped in html tags used for highlighting}
\usage{
cass_highlight(string, pattern, ignore_case = TRUE)
}
\arguments{
\item{string}{A character vector.}

\item{ignore_case}{Defaults to TRUE.}

\item{param}{Pattern to match.}
}
\description{
Takes a character vector and returns it with matches of pattern wrapped in html tags used for highlighting
}
\examples{
cass_highlight(
  string = c(
    "The R Foundation for Statistical Computing",
    "R is free software and comes with ABSOLUTELY NO WARRANTY"
  ),
  pattern = "foundation|software|warranty"
)
}
